% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_survey_results2.R
\name{analyse_survey_results2}
\alias{analyse_survey_results2}
\title{Analyse Survey Results 2}
\usage{
analyse_survey_results2(
  df,
  file_path = NULL,
  sample_design,
  aggregation = NULL,
  strata = NULL,
  cluster = NULL,
  svy_weights = NULL,
  proportions = NULL,
  means = NULL,
  ratios_rates.numerators = NULL,
  ratios_rates.denominators = NULL,
  ratios_rates.multiplier = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe with survey data}

\item{file_path}{Optional input of a character file path which to save an excel sheet of the results.}

\item{sample_design}{Input to specify the survey design. Options include 'srs' for simple random sampling, 'two_stage_cluster',
'two_stage_stratified', and 'two_stage_stratified_cluster'.}

\item{aggregation}{Optional input of character value specifying by which variable you want to aggregate or group the results.}

\item{strata}{Optional input for character value specifying the column name of the strata variable.}

\item{cluster}{Optional input for character value specifying the column name of the cluster variable.}

\item{svy_weights}{Optional input for character value specifying the column name of the svy_weights variable.}

\item{proportions}{Optional input for a character vector specifying all the column names to analyse as proportions.
If categorical variables are included, then dummy variables will be created for each option.}

\item{means}{Optional input for a character vector specifying all the column names to analyse as means}

\item{ratios_rates.numerators}{Optional input for character vector specifying all the column names of numerators for rates or ratios to analyse.}

\item{ratios_rates.denominators}{Optional input for character vector specifying all the column names of denominators for rates or ratios to analyse.}

\item{ratios_rates.multiplier}{Optional input for numeric value specifying a number to multiply the rate results by.}
}
\value{
Returns a dataframe of aggregated results, with point estimates and 95\% confidence intervals within the same cell value,
by aggregation specified or overall.
}
\description{
Generalizable function to analyse survey data for proportions, means, rates or ratios. Wraps srvyr functions.
will also return the survey design object for further analysis
}
\examples{
\dontrun{analyse_survey_results(df = proc_mortality1, ratios_tates.numerators = c(deaths, under5_deaths),
ratios_rates.denominators = c(persontime, under5_persontime), ratios_rates_multiplier = 10000}
}
