% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_cleaning_log.R
\name{append_cleaning_log}
\alias{append_cleaning_log}
\title{Append Cleaning Log}
\usage{
append_cleaning_log(old_cl, new_cl)
}
\arguments{
\item{old_cl}{Inputs a dataframe of the current, working cleaning log file.}

\item{new_cl}{Inputs a dataframe of a new, automated cleaning log file, to be added to the working cleaning log.}
}
\value{
Returns a combined cleaning log file.
}
\description{
Function to read a new and old REACH style cleaning log, and will return a combined cleaning log of only unique issues.
It will preferentially keep cleaning log lines from the old, working cleaning log file, adding in any new issues from
the new cleaning log file.
}
\examples{
\dontrun{append_deletion_log(old_cl = myworkingcl, new_cl = mynewcl)}
}
