% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_map.R
\name{apply_map}
\alias{apply_map}
\title{Apply Map}
\usage{
apply_map(df, map)
}
\arguments{
\item{df}{A dataframe object to recode using a map dataframe.}

\item{map}{A map dataframe matching new and old values for columns to recode.}
}
\value{
Returns a recoded dataframe in-line with the map file.
}
\description{
Apply Map
}
