% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_apart_cis.R
\name{break_apart_cis}
\alias{break_apart_cis}
\title{Break Apart CIs}
\usage{
break_apart_cis(df, result_cols)
}
\arguments{
\item{df}{Inputs a dataframe that has been returned from healthyr::analyse_survey_results.}

\item{result_cols}{Inputs a character vector of the results column names.}
}
\value{
Returns a dataframe that has reshaped long
}
\description{
Helper function to facilitate reformatting results from healthyr::analyse_survey_results
so they can be more easily used to plot results. It is intended to be called within several of the healthyr
plotting functions, and not genererally used directly.
}
\examples{
\dontrun{break_apart_cis(df, result_cols = c("result_col1", "result_col2", ...))}
}
