% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nut_health_indicators.R
\name{calculate_nut_health_indicators}
\alias{calculate_nut_health_indicators}
\title{Calculate Nutrition Health Composite Indicators and Indices}
\usage{
calculate_nut_health_indicators(
  df,
  monthly_expenditures = NULL,
  period_expenditures = NULL,
  num_period_months = NULL,
  use_flags = NULL
)
}
\arguments{
\item{df}{Inputs a data frame that has already been processed and checked by the format_nut_health_indicators and reformat_nut_health_indicators functions.}

\item{monthly_expenditures}{Inputs a character vector of column names for columns of various household expenses in the previous 30 days.}

\item{period_expenditures}{Inputs a character vector of column names for columns of various household expenses in a previous, recall period, specified in the num_period_months paramater.}

\item{num_period_months}{Inputs a whole integer for the number of months of the period expenditures reported.}

\item{use_flags}{Optional input to specify if IYCF flags should be automatically blanked in the numerator.}
}
\value{
Returns a data frame with additional columns for new nutrition and health indicators.
}
\description{
Function to create new columns for calculated nutrition and health indicators. This function is intended to be a helper function
and to be called from within the format_nut_health_indicators function after column names and values have been standardized. This
function will calculate anthropometric indicis, Washington Group Short Set indicators, infant and young child feeding indicators,
food security outcome indicators, and a few other MSNA relevant health indicators.
}
\examples{
\dontrun{calculate_nut_health_indicators(df)}

#' @importFrom rlang .data
}
