% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plausibility_report.R
\name{calculate_plausibility_report}
\alias{calculate_plausibility_report}
\title{Calculate Plausibility Report}
\usage{
calculate_plausibility_report(df)
}
\arguments{
\item{df}{Inputs a dataframe that has been created from one of the create quality report functions.}
}
\value{
Returns a dataframe with additional columns for plausibility penalty scores and classifications,
depending on the technical indicators that were in the original df.
}
\description{
Calculates an approximation of the standard ENA plausibility scores and classification for anthropometric data.
Individually assess each key parameter, and if all parameters are present it will additionally create the
plausibility score and classification.
}
\examples{
\dontrun{calculate_plausibility_report(df)}
}
