% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_poisson_pvalues.R
\name{calculate_poisson_pvalues}
\alias{calculate_poisson_pvalues}
\title{Calculate Poisson Values}
\usage{
calculate_poisson_pvalues(df, strata = NULL, cluster, case)
}
\arguments{
\item{df}{Inputs a dataframe.}

\item{strata}{Optional input of a character value specifying the strata or groupings to test whether the data
is different than poisson distribution.}

\item{cluster}{Input of a character value specifying the column name for cluster id.}

\item{case}{Input of a character value specifying the column name for the cases of interest.}
}
\value{
Returns a data frame with two columns for strata names and poisson p-values.
}
\description{
Helper function checking how closely the distribution of cases across
clusters follows a poisson distribution. Returns a p-value with the null
hypothesis being that the distribution follows a poisson distribution. Dependency
with the nipnTK package, using the greensIndex function.
}
\examples{
\dontrun{calculate_poisson_pvalues(df, strata = "admin2", cluster = "cluster_id",
case = "gam_wfhz_noflag")}

}
