% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_map_applicable.R
\name{check_map_applicable}
\alias{check_map_applicable}
\title{Check Map Applicable}
\usage{
check_map_applicable(df, map)
}
\arguments{
\item{df}{A dataframe dataset to check against the map dataframe.}

\item{map}{A map dataframe to check against the dataset.}
}
\value{
Returns a Boolean TRUE if the map is applicable for the given dataset.
Otherwise will throw an error and stop.
}
\description{
Check if the map file is applicable for use with the associated dataset.
If not applicable, will throw an error explaining why.
If applicable, will return TRUE
}
