% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_which_analysis_vars.R
\name{check_which_analysis_vars}
\alias{check_which_analysis_vars}
\title{Check Which Analysis Variables}
\usage{
check_which_analysis_vars(df, anthro_grouping_var = NULL)
}
\arguments{
\item{df}{Inputs a dataframe of a standardized anthropometric, FSL, mortality or WG-SS dataset}

\item{anthro_grouping_var}{Inputs an optional character value specifying the grouping by which to make SMART flags}
}
\value{
Returns a vector with values specifying different technical indicators.
}
\description{
A helper function to identify which standard indicators can be checked with
the dashboard based on the data frame input.
}
\examples{
\dontrun{}
}
