% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_helper.R
\name{cleaning_log_helper}
\alias{cleaning_log_helper}
\title{Cleaning Log Helper}
\usage{
cleaning_log_helper(df, uuid, flag, cols, description = NULL)
}
\arguments{
\item{df}{Inputs a dataframe with flags, and selection of columns with values to clean if record is flagged}

\item{uuid}{Inputs a character value of the uuid column name.}

\item{flag}{Inputs a character value of the column specifying flagged records for a single type of issue.
Flag columns must be coded as a 1 for flagged, or 0 if no flag or issue.}

\item{cols}{Inputs a character vector of column names that should be included in the exported cleaning log.}

\item{description}{Inputs a character, or brief custom description of the type of flag. To be included in the cleaning log.}
}
\value{
Returns a REACH style cleaning log listing all flags of the specified type, and variables/values to be checked or cleaned
because of that flag for all flagged records.
}
\description{
Function to facilitate creation of cleaning log segments for single types of flags.
}
\examples{
\dontrun{cl <- cleaning_log_helper(df = df, uuid = "uuid_col", flag = "flag_wgss_age",
cols = c("age_years", wgss_vars), description = "Washington Group Short Set values
are present for a person under-5 years of age. This indicator set is not applicable for
this age group.")}
}
