% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_log_helper_others.R
\name{cleaning_log_helper_others}
\alias{cleaning_log_helper_others}
\title{Cleaning Log Helper for Other Variables}
\usage{
cleaning_log_helper_others(df, uuid)
}
\arguments{
\item{df}{Inputs the dataframe. Must have columns ending in _other.}

\item{uuid}{Inputs a character value of the column with UUID, or other household identifier for cleaning purposes.}
}
\value{
Returns a REACH style cleaning log with all the 'other' responses that need to be checked.
}
\description{
Helper function for the create cleaning log flags function in healthyr. This will create
cleaning log records for each other reported value in the dataset, for all columns ending
in "_other".
}
\examples{
\dontrun{cleaning_log_helper_others(df = mydata, uuid = "_uuid")}
}
