% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette-color-blind.R
\name{color_blind}
\alias{color_blind}
\title{Provide Colorblind Compliant Colors}
\usage{
color_blind()
}
\value{
A vector of 8 Hex RGB definitions.
}
\description{
8 Hex RGB color definitions suitable for charts for colorblind people.
}
\details{
This function is used in others in order to help render plots for those that
are color blind.
}
\examples{
color_blind()

}
\seealso{
Other Color Blind: 
\code{\link{hr_scale_color_colorblind}()},
\code{\link{hr_scale_fill_colorblind}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Color Blind}
