% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anthro_quality_report.R
\name{create_anthro_quality_report}
\alias{create_anthro_quality_report}
\title{Create Anthro Quality Report}
\usage{
create_anthro_quality_report(df, grouping = NULL, file_path = NULL)
}
\arguments{
\item{df}{Inputs a dataframe that has been standardized/formatted y the format_nut_health_indicators function.}

\item{grouping}{Inputs an optional character value as a grouping var. If included, results will be grouped by this variable}

\item{file_path}{Inputs an optional character value specifying a file location and name to save a copy of the output.}
}
\value{
Returns a dataframe with a results table for the anthropometric indicators.
}
\description{
Generates a summary report of anthropometric data including summary results, key quality indicators,
and plausability report results.
}
\examples{
\dontrun{create_anthro_quality_report(df = myanthrodata, index = "wfhz")}
}
