% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fsl_quality_report.R
\name{create_fsl_quality_report}
\alias{create_fsl_quality_report}
\title{Create FSL Quality Report}
\usage{
create_fsl_quality_report(df, grouping = NULL, file_path = NULL)
}
\arguments{
\item{df}{Inputs a data frame that includes FSL outcome data that has already been
processed and standardized by the format_nut_health_indicators function.}

\item{grouping}{Inputs an optional character value specifying a column by which to
group or aggregate the results by.}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}
}
\value{
Returns a dataframe of a summary table of the quality indicators.
}
\description{
This function summarizes key indicators for food security outcome indicators to help
the analyst assess the overall FSL data quality. This summary is intended for FSL
outcome indicators only, focusing on FCS, HHS, rCSI, HDDS, and LCS indicators. This
function also calls a helper function 'healthyr::calculate_plausiblity_report' to assign
penalty scores and recommend an overall quality level for the FSL data.
}
\examples{
\dontrun{create_fsl_report(df = proc_fsl1, grouping = "enum",
file_path = "folder/qualityreport.xlsx")}
}
