% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mortality_quality_report.R
\name{create_mortality_quality_report}
\alias{create_mortality_quality_report}
\title{Create Mortality Quality Report}
\usage{
create_mortality_quality_report(
  df,
  grouping = NULL,
  file_path = NULL,
  exp_sex_ratio = NULL,
  exp_ratio_0_4 = NULL,
  exp_ratio_2_5 = NULL,
  exp_ratio_5_10 = NULL,
  exp_hh_size = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe that has been standardized by the format_mortality_current_census function.}

\item{grouping}{A character value specifying the column name by which you want to generate disaggregated results for
for the mortality quality summary.}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}

\item{exp_sex_ratio}{Inputs a numeric value specifying the expected sex ratio in the population.
If no value is given, a default value of 1:1 is used.}

\item{exp_ratio_0_4}{Inputs a numeric value specifying the expected ratio of people under-5 years
of age : people greater than or equal to 5 years of age in the population. If no value is given, a
default of 0.25 is used, which is about 20\% of the population is under-5. It is highly recommended to
use an actual value specific to your context as this default ratio is likely not specific enough.}

\item{exp_ratio_2_5}{Inputs a numeric value specifying the expected ratio of people under-5 years
of age : people greater than or equal to 5 years of age in the population. If no value is given, a
default of 0.7 is used, which is about 41\% are children under-2 years of age out of the under-5 population. It is highly recommended to
use an actual value specific to your context as this default ratio is likely not specific enough.}

\item{exp_ratio_5_10}{Inputs a numeric value specifying the expected ratio of people under-5 years
of age : people greater than or equal to 5 years of age in the population. If no value is given, a
default of 1.1 is used, which is about 52\% are children under-5 years of age out of the under-10 population. It is highly recommended to
use an actual value specific to your context as this default ratio is likely not specific enough.}

\item{exp_hh_size}{Inputs a numeric value specifying the expected average household size. If no value is given, a
default of 5 is used It is highly recommended to use an actual value specific to your context as this default ratio
is likely not specific enough.}
}
\value{
Returns a dataframe with summary indicators on the mortality data.
}
\description{
This function generates a summary table of the mortality data, either at an overall level or disaggregated by
a specified grouping, such as team_id, cluster, or other.
}
\examples{
\dontrun{create_mortality_quality_report(df)}

}
