% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_value_map.R
\name{create_value_map}
\alias{create_value_map}
\title{Create Value Map
Creates a map of the categorical/factor variables in the dataset.}
\usage{
create_value_map(df)
}
\arguments{
\item{df}{A dataframe object.}
}
\value{
Returns a dataframe with mapped variables and values in a condensed
format, and also adds a "new_value" column for recoding with other functions.
}
\description{
Create Value Map
Creates a map of the categorical/factor variables in the dataset.
}
