% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-files.R
\docType{data}
\name{dx_cc_mapping}
\alias{dx_cc_mapping}
\title{Diagnosis to Condition Code Mapping file}
\format{
A data frame with 86852 rows and 5 variables
}
\usage{
data(dx_cc_mapping)
}
\description{
Diagnosis to Condition Code Mapping file
}
\concept{Data Files

A dataset containing the Diagnosis Code to AHRQ Condition Code Mapping.}
\keyword{datasets}
