% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_ae_acute_trust_attribution}
\alias{england_ae_acute_trust_attribution}
\title{England Attribution from Type 3 A&E Providers to Acute Trusts (updated June 2022)}
\format{
A data frame with 218 rows and 3 variables:
\describe{
\item{nhs_trust22_code_all}{NHS Trust (organisational) code}
\item{nhs_trust22_code_acute}{NHS Trust (organisational) code - Acute Trusts only}
\item{proportion_attendances_attributed_to_acute_trust}{Percentage of A&E attendances attributed to the Acute Trust}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_ae_acute_trust_attribution
}
\description{
A dataset mapping proportion of A&E attendances attributed to Acute Trusts from Type 3 A&E Providers,
including Minor Injury Units and Walk-in Centres
}
\keyword{datasets}
