% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_criteria_to_reside}
\alias{england_criteria_to_reside}
\title{Hosptial Discharge Data - Criteria to Reside}
\format{
A data frame with 29,524 rows and 3 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust (organisational) code}
\item{date}{Date}
\item{do_not_meet_criteria_to_reside}{Number of patients who no longer meet the criteria to reside}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_criteria_to_reside
}
\description{
A dataset containing NHS Trust level hopsital discharge data on how many
patients do not meet criteria to reside each day.
}
\keyword{datasets}
