% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_delayed_discharge_reasons}
\alias{england_delayed_discharge_reasons}
\title{Disability by LTLA for England, 2021}
\format{
A data frame with 45,220 rows and 5 variables:
\describe{
\item{nhs_trust22_name}{NHS Trust name}
\item{nhs_trust22_code}{NHS Trust code}
\item{date}{Date in the format year-month-day}
\item{delay_reason}{The reason for the delayed discharge}
\item{delay_count}{The number of delays}
...
}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/discharge-delays/acute-discharge-situation-report/}
}
\usage{
england_delayed_discharge_reasons
}
\description{
A dataset containing LTLA data on numbers disabled under the Equality Act.
}
\keyword{datasets}
