% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_disability_21}
\alias{england_disability_21}
\title{Disability by LTLA England, 2021}
\format{
A data frame with 662 rows and 5 variables:
\describe{
\item{ltla21_code}{LTLA code}
\item{total_residents}{Total residents in the local authority}
\item{disability level}{Disabled or Not disabled under the Equality Act}
\item{n}{number of people}
\item{prop}{proportion of people}
...
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
england_disability_21
}
\description{
A dataset containing LTLA data on numbers disabled under the Equality Act.
}
\keyword{datasets}
