% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_health_index_ics}
\alias{england_health_index_ics}
\title{England Health Index (ONS)}
\format{
A data frame with 294 rows and 7 variables:
\describe{
\item{icb22_code}{2022 Integrated Care Board code}
\item{icb22_name}{2022 Integrated Care Board name}
\item{year}{Year}
\item{overall_score}{The overall health index score}
\item{healthy_people_domain_score}{Score for the Healthy People domain}
\item{healthy_lives_domain_score}{Score for the Healthy Lives domain}
\item{healthy_places_domain_score}{Score for the Healthy Places domain}
...
}
}
\source{
\url{https://www.ons.gov.uk/}
}
\usage{
england_health_index_ics
}
\description{
A dataset containing overall health index scores for the English Integrated
Care Boards from 2015-2021.
}
\keyword{datasets}
