% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_health_index_ics_indicators}
\alias{england_health_index_ics_indicators}
\title{England Health Index (ONS) - subdomains}
\format{
A data frame with 4,116 rows and 5 variables:
\describe{
\item{icb22_code}{2022 Integrated Care Board code}
\item{icb22_name}{2022 Integrated Care Board name}
\item{year}{Year}
\item{sub_domain}{The name of the subdomain}
\item{value}{The value of the indicator}
...
}

A data frame with 16,464 rows and 5 variables:
\describe{
\item{icb22_code}{2022 Integrated Care Board code}
\item{icb22_name}{2022 Integrated Care Board name}
\item{year}{Year}
\item{indicator}{The name of the indicator}
\item{value}{The value of the indicator}
...
}
}
\source{
\url{https://www.ons.gov.uk/}

\url{https://www.ons.gov.uk/}
}
\usage{
england_health_index_ics_indicators

england_health_index_ics_indicators
}
\description{
A dataset containing subdomain health index scores for the English Integrated
Care Boards from 2015-2021.

A dataset containing underlying indicators for the ONS Health Index for
English Integrated Care Boards from 2015-2021.
}
\keyword{datasets}
