% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_health_index_indicators}
\alias{england_health_index_indicators}
\title{England Health Index (ONS) - underlying indicators}
\format{
A data frame with 120,344 rows and 4 variables:
\describe{
\item{ltla21_code}{2021 Lower tier local authority code}
\item{year}{Year}
\item{indicator}{The name of the indicator}
\item{value}{The value of the indicator}
...
}
}
\source{
\url{https://www.ons.gov.uk/}
}
\usage{
england_health_index_indicators
}
\description{
A dataset containing underlying indicators for the ONS Health Index for
English lower tier local authorities from 2015-2021.
}
\keyword{datasets}
