% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_health_index_subdomains}
\alias{england_health_index_subdomains}
\title{England Health Index (ONS) - subdomains}
\format{
A data frame with 2,149 rows and 16 variables:
\describe{
\item{ltla21_code}{2021 Lower tier local authority code}
\item{year}{Year}
\item{Difficulties in daily life}{Difficulties in daily life - Health People domain}
\item{Mental health}{Mental health - Health People domain}
\item{Mortality}{Mortality - Health People domain}
\item{Personal well-being}{Personal well-being - Health People domain}
\item{Physical health conditions}{Physical health conditions - Health People domain}
\item{Behavioural risk factors}{Behavioural risk factors - Health Lives domain}
\item{Children and young people}{Children and young people - Health Lives domain}
\item{Physiological risk factors}{Physiological risk factors - Health Lives domain}
\item{Protective measures}{Protective measures - Health Lives domain}
\item{Access to green space}{Access to green space - Health Places domain}
\item{Access to services}{Access to services - Health Places domain}
\item{Crime}{Crime - Health Places domain}
\item{Economic and working conditions}{Economic and working conditions - Health Places domain}
\item{Living conditions}{Living conditions - Health Places domain}
...
}
}
\source{
\url{https://www.ons.gov.uk/}
}
\usage{
england_health_index_subdomains
}
\description{
A dataset containing overall health index scores for the English lower tier
local authorities from 2015-2021.
}
\keyword{datasets}
