% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_iapt}
\alias{england_iapt}
\title{Physchological Therpaies - IAPT}
\format{
A data frame with 52,932 rows and 4 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust (organisational) code}
\item{date}{Date}
\item{name}{Name of the variable}
\item{value}{Value of the variable}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_iapt
}
\description{
A dataset containing NHS Trust level IAPT data.
}
\keyword{datasets}
