% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_icb_accidents_emergency}
\alias{england_icb_accidents_emergency}
\title{England Accident and Emergency Admissions by Integrated Care Board  - (2022-23)}
\format{
A data frame with 504 rows and 8 variables:
\describe{
\item{code}{Integrated Care Board code}
\item{total_attendances}{The total number of patients in an A&E service seeking medical attention}
\item{attendances_over_4hours}{The number of patients seeking medical attention that spend over 4 hours from arrival to admission, transfer or discharge}
\item{total_emergency_admissions}{The total number of admissions to a hospital bed as an emergency}
\item{emergency_admissions_over_4hours}{The number of patients seeking admission to a hospital bed as an emergency that spend over 4 hours from decision to admit to admission}
\item{pct_attendance_over_4hours}{The percentage of attendance that spend over 4 hours from arrival to admission, transfer of discharge}
\item{date}{Date}
\item{pct_emergency_admissions_over_4hours}{The percentage of emergency admissions that spend over 4 hours from decision to admit to admission}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_icb_accidents_emergency
}
\description{
A dataset containing England Accident and Emergency attendances and Emergency Admissions at Integrated Care Board level, by month
}
\keyword{datasets}
