% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_icb_criteria_to_reside}
\alias{england_icb_criteria_to_reside}
\title{Hospital ICB Discharge Data - Criteria to Reside}
\format{
A data frame with 24,239 rows and 3 variables:
\describe{
\item{icb22_code}{Integrated Care Board code}
\item{date}{Date}
\item{do_not_meet_criteria_to_reside}{Number of patients who no longer meet the criteria to reside}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_icb_criteria_to_reside
}
\description{
A dataset containing NHS Integrated Care Board discharge data on how many
patients do not meet criteria to reside each day.
}
\keyword{datasets}
