% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_legal_partnership_21}
\alias{england_legal_partnership_21}
\title{Legal Partnership Status, England and Wales, 2021}
\format{
A data frame with 5,627 rows and 5 variables:
\describe{
\item{ltla21_code}{LTLA code}
\item{total_population}{Population in the LTLA}
\item{legal_partnership}{Type of legal partnership}
\item{n}{Number of people}
\item{prop}{Percentage}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
england_legal_partnership_21
}
\description{
A dataset containing legal partnership status data at LTLA level
}
\keyword{datasets}
