% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_ambulance_quality_indicators}
\alias{england_nhs_ambulance_quality_indicators}
\title{Ambulance Quality Indicators Data 2022-23}
\format{
A tibble data frame with 256 rows and 8 variables:
\describe{
\item{trust_code}{Abbreviation for trust name}
\item{trust_name}{Trust Name}
\item{Ambulance Service}{Name of Amblunce service}
\item{Incident Count}{The counts of incidence in each NHS Trust}
\item{Total_Response_Time_h}{Total response time in hours}
\item{Mean_Response_Time_min_sec}{Total response time in minutes and seconds}
\item{90th_Centile_Response_Time_min_sec}{The 90\% Total response time in minutes and seconds}
\item{Category}{Category}

}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/ambulance-quality-indicators/ambulance-quality-indicators-data-2022-23/}
}
\usage{
england_nhs_ambulance_quality_indicators
}
\description{
This data set contains Ambulance System Indicators (AmbSYS)  the incident counts,
and their response time.
}
\keyword{datasets}
