% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_bed_occupancy}
\alias{england_nhs_bed_occupancy}
\title{Bed Availability and Occupancy}
\format{
A tibble data frame with 212 rows and 22 variables:
\describe{
\item{\% trust_code}{Abbreviation for trust name}
\item{\% trust_name}{Trust Name}
\item{\% Total_Day_Beds_Occupied}{Beds occupied durring the day}
\item{General_Acute_Day_Beds_Occupied}{Beds Occupied durring the day in General_Acute unit}
\item{Learning Disabilities Day Beds Occupied}{Beds Occupied durring the day in Learning Disabilities unit}
\item{Maternity Night day Occupied}{Beds Occupied durring the day in the Maternity unit}
\item{Mental Illness Day Beds Occupied}{Beds Occupied durring the day in Mental Illness unit}
\item{\% Total Day Beds Occupied }{Total Beds Occupied durring the day }
\item{\% General_Acute_Day_Beds_Occupied}{Beds Occupied durring the day in General_Acute unit}
\item{\% Learning Disabilities Day Beds Occupied}{Beds Occupied durring the day in Learning Disabilities unit}
\item{\% Maternity day Occupied}{Beds Occupied durring the day in the Maternity unit}
\item{\% Mental Illness Day Beds Occupied}{Beds Occupied durring the day in Mental Illness unit}
\item{\% Total Night Beds Occupied}{Total Night Beds Occupied}
\item{General Acute Night Beds Occupied}{Beds Occupied durring the night in General_Acute_unit}
\item{Learning Disabilities Night Beds Occupied}{Beds Occupied durring the night in Learning Disabilities unit}
\item{Maternity Night Beds Occupied}{Beds Occupied durring the night in the Maternity unit}
\item{Mental Illness Night Beds Occupied}{Mental Illness Night Beds Occupied}
\item{\% Total Night Beds Occupied}{Total Night Beds Occupied}
\item{\% General Acute Night Beds Occupied}{Beds Occupied durring the night in General_Acute_unit}
\item{\% Learning Disabilities Night Beds Occupied}{Beds Occupied durring the night in Learning Disabilities unit}
\item{\% Maternity Night Beds Occupied}{Beds Occupied durring the night in the Maternity unit}
\item{\% Mental Illness Night Beds Occupied}{Beds Occupied durring the night in Mental Illness unit}

}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/ambulance-quality-indicators/ambulance-quality-indicators-data-2022-23/}
}
\usage{
england_nhs_bed_occupancy
}
\description{
This data set contains the  number of available bed days & night.
It gives the  number of occupied bed days by consultant main specialty.
}
\keyword{datasets}
