% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_cancer_wait_times}
\alias{england_nhs_cancer_wait_times}
\title{Cancer Waiting Times}
\format{
A tibble data frame with 1400 rows and 6 variables:
\describe{
\item{\% trust_code}{Abbreviation for trust name}
\item{\% trust_name}{Trust Name}
\item{Standard}{ST}
\item{Total Treated}{Total number of patients treated}
\item{Within Standard}{WS}
\item{Breaches}{Breaches}
...

}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/cancer-waiting-times/}
}
\usage{
england_nhs_cancer_wait_times
}
\description{
This data set contains the waiting times of people referred by GP with suspected cancer or
breast symptoms and those subsequently diagnosed with and treated for cancer by the NHS in England.
}
\keyword{datasets}
