% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_diagnostic_wait_times}
\alias{england_nhs_diagnostic_wait_times}
\title{Cancer Waiting Times}
\format{
A tibble data frame with 212 rows and 7 variables:
\describe{
\item{\% trust_code}{Abbreviation for trust name}
\item{\% trust_name}{Trust Name}
\item{Waiting List Total}{Waiting list of patients}
\item{Waiting 6+ weeks}{Waiting list of patients for 6 Weeks plus}
\item{Waiting 13+ weeks}{Waiting list of patients for 6 Weeks plus}
\item{\% Waiting 6+ weeks}{Total number Waiting list of patients for 6 Weeks plus}
\item{\% Waiting 13+ weeks}{Total number Waiting list of patients for 13 Weeks plus}
...

}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/diagnostics-waiting-times-and-activity/monthly-diagnostics-waiting-times-and-activity/monthly-diagnostics-data-2022-23/}
}
\usage{
england_nhs_diagnostic_wait_times
}
\description{
This data set contains the waiting times of people referred by GP with suspected cancer or
breast symptoms and those subsequently diagnosed with and treated for cancer by the NHS in England.
}
\keyword{datasets}
