% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_iapt}
\alias{england_nhs_iapt}
\title{Improving Access to Psychological Therapies (IAPT)}
\format{
A tibble data frame with 42 rows and 2 variables:
\describe{
\item{stp_code_short}{stp_code_short   }
\item{Referrals that finished a course of treatment in the month waiting 126 days or less for first treatment}{Referrals that finished a course of treatment in the month waiting 126 days or less for first treatment}

....

}
}
\source{
\url{https://digital.nhs.uk/data-and-information/publications/statistical/psychological-therapies-report-on-the-use-of-iapt-services/june-2021-final-including-reports-on-the-iapt-pilots-and-quarter-1-data-2021-22}
}
\usage{
england_nhs_iapt
}
\description{
This data set contains the referrals patients that have finished a course of treatment in the
month waiting 126 days or less for first treatment.
}
\keyword{datasets}
