% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_outpatient_referrals}
\alias{england_nhs_outpatient_referrals}
\title{Outpatient Referrals Data}
\format{
A tibble data frame with 212 rows and 6 variables:
\describe{
\item{Trust_code}{Abbreviation for trust name}
\item{Trust_name}{Trust Name}
\item{GP Referrals Made (All)}{The total number of refferals made in all units by the GP}
\item{Other Referrals Made (All)}{The total number of refferals made in all units}
\item{GP Referrals Made (Specific Acute)}{The number of refferals made in Specific Acute unit by the GP}
\item{Other Referrals Made (Specific Acute)}{The number of refferals made in Specific Acute unit}
...

}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/outpatient-referrals/mrr-data/}
}
\usage{
england_nhs_outpatient_referrals
}
\description{
This data set contains data relating to GP and other referrals for first consultant-led
outpatient appointments focusing on all treatment functions and specific acute treatment functions.
}
\keyword{datasets}
