% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{england_nhs_referral_treatment_waiting_times}
\alias{england_nhs_referral_treatment_waiting_times}
\title{Consultant-led Referral to Treatment Waiting Times}
\format{
A tibble data frame with 824 rows and 8 variables:
\describe{
\item{Trust_code}{Abbreviation for trust name}
\item{Trust_name}{Trust Name}
\item{Referral Treatment Type}{Waiting list of patients}
\item{Waiting 52+ Weeks}{Waiting list of patients for 52+ Weeks plus}
\item{Waiting 18+ weeks}{Waiting list of patients for 18+ Weeks plus}
\item{Total waiting list}{Total number of patients on the Waiting list}
\item{\% Waiting 18+ weeks}{Percentage of the Waiting list of patients for 18+ Weeks plus}
\item{\% Waiting 52+ weeks}{Percentage of theWaiting list of patients for 52+ Weeks plus}
...

}
}
\source{
\url{https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/rtt-data-2022-23/}
}
\usage{
england_nhs_referral_treatment_waiting_times
}
\description{
This data set contains contains information on Consultant-led Referral To Treatment (RTT) waiting times,
which monitor the length of time from referral through to elective treatment.
}
\keyword{datasets}
