% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_rtt_flow_stp}
\alias{england_rtt_flow_stp}
\title{Consultant-led Referral to Treatment (RTT) pathway flow by ICS/STP}
\format{
A data frame with 2,478 rows and 7 variables:
\describe{
\item{stp20_nhs_code}{NHS code for the ICS/STP}
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{completed}{Number of pathways completed}
\item{incomplete}{Number of incomplete pathways}
\item{new}{Number of new pathways}
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_rtt_flow_stp
}
\description{
A dataset containing ICS/STP-level RTT data spanning from Apr 2020 to Feb
2025.
}
\keyword{datasets}
