% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_rtt_region}
\alias{england_rtt_region}
\title{Consultant-led Referral to Treatment (RTT) by NHS Region}
\format{
A data frame with 6,804 rows and 7 variables:
\describe{
\item{nhs_region20_name}{Name of NHS England region}
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{treatment}{Treatment function name}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
\item{waits_over_52_weeks}{Total waiting > 52 weeks}
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_rtt_region
}
\description{
A dataset containing ICS/STP-level RTT data.
}
\keyword{datasets}
