% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_rtt_stp}
\alias{england_rtt_stp}
\title{Consultant-led Referral to Treatment (RTT) by ICS/STP}
\format{
A data frame with 55,73 rows and 7 variables:
\describe{
\item{stp20_nhs_code}{NHS code for the ICS/STP}
\item{stp20_name}{Name of the ICS/STP}
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{treatment}{Treatment function name}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
\item{waits_over_52_weeks}{Total waiting > 52 weeks}
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_rtt_stp
}
\description{
A dataset containing ICS/STP-level RTT data spanning from Apr 2020 to Feb
2025.
}
\keyword{datasets}
