% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_rtt_trust}
\alias{england_rtt_trust}
\title{Consultant-led Referral to Treatment (RTT) by NHS Trust}
\format{
A data frame with 11,676 rows and 7 variables:
\describe{
\item{nhs_trust22_code}{Code of NHS England trust}
\item{nhs_trust22_name}{Name of NHS England trust}
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{treatment}{Treatment function name}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
\item{waits_over_52_weeks}{Total waiting > 52 weeks}
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_rtt_trust
}
\description{
A dataset containing ICS/STP-level RTT data spanning from Jan 2019 to Feb
2025.
}
\keyword{datasets}
