% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_sexual_orientation_21}
\alias{england_sexual_orientation_21}
\title{Sexual orientation, England and Wales, 2021}
\format{
A data frame with 1,655 rows and 5 variables:
\describe{
\item{ltla21_code}{LTLA code}
\item{total_residents}{Residents in the LTLA}
\item{sexual_orientation}{Sexual orientation}
\item{n}{Number of people}
\item{prop}{Percentage}
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
england_sexual_orientation_21
}
\description{
A dataset containing sexual orientation data at LTLA level
}
\keyword{datasets}
