% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_sicb_iapt}
\alias{england_sicb_iapt}
\title{Physchological Therpaies - IAPT}
\format{
A data frame with 8,162 rows and 4 variables:
\describe{
\item{sicbl22_code}{Sub Integrated Care Board code}
\item{date}{Date}
\item{name}{Name of the variable}
\item{value}{Value of the variable}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_sicb_iapt
}
\description{
A dataset containing NHS Sub Integrated Care Board (formerly CCGs) data.
}
\keyword{datasets}
