% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_trust_accidents_emergency}
\alias{england_trust_accidents_emergency}
\title{England Accident and Emergency Admissions by Trust - (2021-25)}
\format{
A data frame with 7,895 rows and 8 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust (organisational) code}
\item{total_attendances}{The total number of patients in an A&E service seeking medical attention}
\item{attendances_over_4hours}{The number of patients seeking medical attention that spend over 4 hours from arrival to admission, transfer or discharge}
\item{total_emergency_admissions}{The total number of admissions to a hospital bed as an emergency}
\item{emergency_admissions_over_4hours}{The number of patients seeking admission to a hospital bed as an emergency that spend over 4 hours from decision to admit to admission}
\item{pct_attendance_over_4hours}{The percentage of attendance that spend over 4 hours from arrival to admission, transfer of discharge}
\item{date}{Date}
\item{pct_emergency_admissions_over_4hours}{The percentage of emergency admissions that spend over 4 hours from decision to admit to admission}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_trust_accidents_emergency
}
\description{
A dataset containing England Accident and Emergency attendances and Emergency Admissions at provider level, by month.
}
\details{
Data notes:
\itemize{
\item Due to a cyber-attack several sites have been unable to provide complete data since August 2022
\item Fourteen trusts are field testing new A&E performance standards and as a result are not required to report attendances over four hours from May 2019
}
}
\keyword{datasets}
