% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_health_records.R
\name{extract_health_records}
\alias{extract_health_records}
\title{Extract Records from health_data object}
\usage{
extract_health_records(health_data)
}
\arguments{
\item{health_data}{output from \code{import_health_data()}}
}
\value{
data.frame
}
\description{
Extract Records from health_data object
}
