% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_anthro_issues.R
\name{flag_anthro_issues}
\alias{flag_anthro_issues}
\title{Flag Anthropometric Issues}
\usage{
flag_anthro_issues(df, grouping = NULL, file_path = NULL)
}
\arguments{
\item{df}{Inputs a data frame with anthropometric data that has already been processed by the format_nut_health_indicators and reformat_nut_health_indicators
functions.}

\item{grouping}{Inputs a character value specifying a column name to indicate the groupings for SMART flag creation.}

\item{file_path}{Inputs an optional character value with the file path for saving an xlsx file of the output}
}
\value{
Returns a data frame with additional columns for WHO and SMART flags
}
\description{
This function creates SMART and WHO flags in an anthropometric dataset, using the mean and sd within the grouping provided for SMART flags.
}
\examples{
flag_anthro_issues(df = proc_anthro1, grouping = "county")

}
