% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_mortality_issues.R
\name{flag_mortality_issues}
\alias{flag_mortality_issues}
\title{Flag Mortality Issues}
\usage{
flag_mortality_issues(df)
}
\arguments{
\item{df}{Inputs a dataframe that has been formatted by the format_mortality_current_census function, which
is intended to reformat data collected using the current census method via the SMART methodology.}
}
\value{
Returns dataframe, the mortality dataset with additional columns flagging records with potential
data quality issues.
}
\description{
To flag individual households which may have unusual demographic and mortality reporting, for purposes of
data cleaning and quality monitoring.
}
\examples{
\dontrun{flag_mortality_issues(proc_mortality1)}

}
