% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_nut_health_issues.R
\name{flag_nut_health_issues}
\alias{flag_nut_health_issues}
\title{Flag Nutrition Health Issues}
\usage{
flag_nut_health_issues(df)
}
\arguments{
\item{df}{Inputs a dataframe which has already been processed by the format_nut_health_indicators, reformat_nut_health_indicators, and
calculate_nut_health_indicators functions. This function is not intended to be run directly, and instead is called automatically
within the format_nut_health_indicators function.}
}
\value{
Returns a dataframe with additional columns of flags/data quality checks for individual records.
}
\description{
Flag Nutrition Health Issues
}
\examples{
\dontrun{flag_nut_health_issues(df)}

}
