% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_anthro_to_ena.R
\name{format_anthro_to_ena}
\alias{format_anthro_to_ena}
\title{Format Anthropometric Data to ENA Format}
\usage{
format_anthro_to_ena(df)
}
\arguments{
\item{df}{Inputs a dataframe that has been already formatted by the format_nut_health_indicators function.}
}
\value{
Returns a dataframe which is formatted so it can be copy-pasted into the anthropometric section of the ENA software.
}
\description{
Format Anthropometric Data to ENA Format
}
\examples{
format_anthro_to_ena(df = proc_anthro1)

}
