% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mortality_to_ena.R
\name{format_mortality_to_ena}
\alias{format_mortality_to_ena}
\title{Reformat Mortaltiy to ENA}
\usage{
format_mortality_to_ena(df, file_path = NULL)
}
\arguments{
\item{df}{Inputs a dataframe which has been standardized already by the format_mortality_current_census function.}

\item{file_path}{Inputs an optional character value for specifying a file path to save the output as an xlsx file.}
}
\value{
Returns a dataframe with wide mortality data, that is each row is a single household, which is formatted for ENA.
The resulting dataframe can be copy-pasted straight into the ENA program for analysis of mortality data.
}
\description{
This function will take a standardized mortality dataset and reshape it for use in the ENA software.
}
\examples{
format_mortality_to_ena(df = proc_mortality1)

}
