% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_muac_dps.R
\name{helper_muac_dps}
\alias{helper_muac_dps}
\title{Helper MUAC DPS}
\usage{
helper_muac_dps(muac)
}
\arguments{
\item{muac}{A vector of MUAC measurements.}
}
\value{
Returns a vector of digit preference scores.
}
\description{
A helper function for returning MUAC digit preference scores within a
dplyr call.
}
\examples{
\dontrun{helper_muac_dps(muac)}

}
