% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette-color-blind.R
\name{hr_scale_color_colorblind}
\alias{hr_scale_color_colorblind}
\title{Provide Colorblind Compliant Colors}
\usage{
hr_scale_color_colorblind(..., theme = "hr")
}
\arguments{
\item{...}{Data passed in from a \code{ggplot} object}

\item{theme}{Right now this is \code{hr} only. Anything else will render an error.}
}
\value{
A \code{gggplot} layer
}
\description{
8 Hex RGB color definitions suitable for charts for colorblind people.
}
\details{
This function is used in others in order to help render plots for those that
are color blind.
}
\seealso{
Other Color Blind: 
\code{\link{color_blind}()},
\code{\link{hr_scale_fill_colorblind}()}
}
\author{
Steven P. Sanderson II, MPH
}
\concept{Color Blind}
