% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_cleaning_log.R
\name{implement_cleaning_log}
\alias{implement_cleaning_log}
\title{Implement Cleaning Log}
\usage{
implement_cleaning_log(df, uuid_col, cleaning_log)
}
\arguments{
\item{df}{Inputs a dataframe of raw data.}

\item{uuid_col}{Inputs a character value specifying the column name for uuid, or
whichever unique record level value which to target cleaning.}

\item{cleaning_log}{Inputs a dataframe that is the REACH-style cleaning log.}
}
\value{
Returns a dataframe of the clean dataset.
}
\description{
Function to read a REACH-style cleaning log and a dataframe and produce a cleaned
dataset based on the cleaning log.
}
\examples{
\dontrun{implement_cleaning_log(df = rawdata, uuid_col = "uuid",
cleaning_log = mycleaninglog)}

}
