% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_deletion_log.R
\name{implement_deletion_log}
\alias{implement_deletion_log}
\title{Implement Deletion Log}
\usage{
implement_deletion_log(df, deletion_log, uuid_col)
}
\arguments{
\item{df}{Inputs the dataframe with records to be deleted.}

\item{deletion_log}{Inputs the REACH style deletion log, indicating which records to be deleted. Must have a
'uuid' named column.}

\item{uuid_col}{Inputs a character value specifying the column name of the dataframe with the unique identifier which matches the
deletion log uuid column.}
}
\value{
Returns a dataframe without the deleted records.
}
\description{
Will delete records from a dataframe according to a REACH style deletion log.
}
\examples{
\dontrun{implement_deletion_log(df = mydata, deletion_log = my_deletion_log, uuid_col = "my_hh_id")}
}
