% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_health_data.R
\name{import_health_data}
\alias{import_health_data}
\title{Import Apple Health data}
\usage{
import_health_data(path, type = "xml")
}
\arguments{
\item{path}{chr path to exported data}

\item{type}{chr type of data; currently only supports 'xml'}
}
\value{
object of classes XMLInternalDocument and XMLAbstractDocument
}
\description{
Import Apple Health data
}
