% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_string_split.R
\name{mid}
\alias{mid}
\title{Use SQL LEFT, MID and RIGHT type functions}
\usage{
mid(text, start_num, num_char)
}
\arguments{
\item{text}{A piece of text/string to be manipulated}

\item{start_num}{What place to start at}

\item{num_char}{How many characters do you want to grab}
}
\description{
perform either left, mid or right substring sql like manipulations
}
\details{
\itemize{
\item You must supply data that you want to manipulate and substrig.
}
}
\examples{

mid("this is some text", 3, 5)

}
