% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhs_criteria_to_reside_22}
\alias{nhs_criteria_to_reside_22}
\title{Hosptial Discharge Data - Criteria to Reside (2022)}
\format{
A data frame with 25,894 rows and 3 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust (organisational) code}
\item{date}{Date}
\item{do_not_meet_criteria_to_reside}{Number of patients who no longer meet the criteria to reside}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
nhs_criteria_to_reside_22
}
\description{
A dataset containing NHS Trust level hopsital discharge data on how many
patients do not meet criteria to reside for each day of 2022.
}
\keyword{datasets}
