% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhs_iapt_22}
\alias{nhs_iapt_22}
\title{Physchological Therpaies - IAPT (2022)}
\format{
A data frame with 6,876 rows and 4 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust (organisational) code}
\item{date}{Date}
\item{name}{Name of the variable}
\item{value}{Value of the variable}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
nhs_iapt_22
}
\description{
A dataset containing NHS Trust level IAPT data.
}
\keyword{datasets}
