% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_bed_occupancy}
\alias{ni_bed_occupancy}
\title{Bed Occupany}
\format{
A data frame with 5 rows and 5 variables:
\describe{
\item{trust18_code}{Health and Social Care Trust code code}
\item{date}{Date}
\item{beds_available_count}{Total number of available beds}
\item{beds_occupied_count}{Total number of occupied beds}
\item{beds_occupied_percentage}{The percentage of occupied beds}
...
}
}
\source{
\url{https://www.health-ni.gov.uk/}
}
\usage{
ni_bed_occupancy
}
\description{
A dataset containing Health and Social Care Trust level bed occupancy data.
}
\keyword{datasets}
