% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_beds}
\alias{ni_beds}
\title{Bed Availability in Northern Ireland}
\format{
A data frame with 13,237 rows and 15 variables:
\describe{
\item{financial_year}{Financial year}
\item{quarter_ending}{Quarter ending date}
\item{HSC}{Health and Social Care Trust locations}
\item{hospital}{Hospital name}
\item{programme_of_care}{Programme of care}
\item{specialty}{Specialty}
\item{total_available_beds}{Total number of available beds}
\item{average_available_beds}{Average number of available beds}
\item{total_occupied_beds}{Total number of occupied beds}
\item{average_occupied_beds}{Average number of occupied beds}
\item{total_inpatients}{Total number of inpatients}
\item{total_day_case}{Total number of day cases}
\item{elective_inpatient}{Number of elective inpatients}
\item{non_elective_inpatient}{Number of non-elective inpatients}
\item{regular_attenders}{Number of regular attenders}
}
}
\source{
\url{https://www.health-ni.gov.uk/}
}
\usage{
ni_beds
}
\description{
A dataset containing hospital statistics on bed availability, bed occupancy and inpatients data.
}
\keyword{datasets}
