% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_rtt_hsct}
\alias{ni_rtt_hsct}
\title{Consultant-led Referral to Treatment (RTT) by Health & Social Care Trust}
\format{
A data frame with 2,706 rows and 7 variables:
\describe{
\item{hsct22_name}{Health & Social Care Trust name}
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{specialty}{Specialty}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
\item{waits_over_52_weeks}{Total waiting > 52 weeks}
\item{total_waits}{Total waits}
}
}
\source{
\url{https://www.health-ni.gov.uk/topics/doh-statistics-and-research/hospital-waiting-times-statistics}
}
\usage{
ni_rtt_hsct
}
\description{
A dataset containing RTT data for Health & Social Care Trusts in NI.
}
\keyword{datasets}
