% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_underdoctored_areas}
\alias{ni_underdoctored_areas}
\title{Underdoctored areas in Northern Ireland}
\format{
A data frame with 6 variables:
\describe{
\item{ltla21_code}{Local Authority code (2021)}
\item{ltla21_name}{Local Authority name (2021)}
\item{date}{year}
\item{total_patients}{Number of registered patients}
\item{total_gp}{Number of GPs}
\item{patients_per_gp}{Number of patients per GP}
}
}
\usage{
ni_underdoctored_areas
}
\description{
A dataset containing the ratio of patients to GP workforce, by Local Authority
}
\keyword{datasets}
