% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_unpaid_care_21}
\alias{ni_unpaid_care_21}
\title{Provision of unpaid care (delayed discharge) in Northern Ireland}
\format{
A data frame with 605 rows and 4 variables:
\describe{
\item{ltla21_name}{Local Government Districts name}
\item{ltla21_code}{Local Government Districts code}
\item{variable}{Age band and count/percentage of unpaid care hours}
\item{value}{Value of the chosen variable}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
ni_unpaid_care_21
}
\description{
A dataset containing unpaid care data in Local Government Districts in NI
from 2021 census.
}
\keyword{datasets}
