% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_waiting_times}
\alias{ni_waiting_times}
\title{Northern Ireland Emergency Care Waiting Times}
\format{
A data frame containing 6 columns.
\describe{
\item{healthboard_name}{Department name}
\item{date}{Date}
\item{total}{Total attendances}
\item{under_4_hours}{Attendances with waiting times under 4 hours}
\item{between_4_12_hours}{Attendances with waiting times between 4 and 12 hours}
\item{over_12_hours}{Attendances with waiting times over 12 hours}
}
}
\source{
\url{https://datavis.nisra.gov.uk/health/ni-emergency-care-waiting-times-data-oct-dec-24.html#table}
}
\usage{
ni_waiting_times
}
\description{
Emergency care waiting times for departments (hospitals) in Norther Ireland Health & Social Care Trusts.
}
\keyword{datasets}
