% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_proxy_distribution.R
\name{plot_age_proxy_distribution}
\alias{plot_age_proxy_distribution}
\title{Plot Age Proxy Distribution}
\usage{
plot_age_proxy_distribution(
  df,
  by_group = NULL,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe that has been processed and standardized
by the healthyr::format_nut_health_indicators function.}

\item{by_group}{Inputs an optional character value specifying the column
by which to facet the resulting plot.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function plots a stacked bar graph showing the distribution of
children under 2 years of age to over 2 years of age.
}
\examples{
\dontrun{plot_age_proxy_distribution(df, by_group = "enum", )}
}
