% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_years_distribution.R
\name{plot_age_years_distribution}
\alias{plot_age_years_distribution}
\title{Plot Age Distribution in Years}
\usage{
plot_age_years_distribution(
  df,
  by_group = NULL,
  min_age = NULL,
  max_age = NULL,
  breaks = NULL,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe that has been processed and standardized by the
healthyr::format_nut_health_indicators and has an age_years column.}

\item{by_group}{Inputs an optional character value specifying a column by which
to facet the plot.}

\item{min_age}{Inputs a numeric value for the minimum age in years to include in the plot.}

\item{max_age}{Inputs a numeric value for the maximum age in years to include in the plot.}

\item{breaks}{Inputs a numeric value for how many years between ticks to plot.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This functions creates a histogram of reported ages in years.
}
\examples{
\dontrun{plot_age_years_distribution(df, by_group = "enum", min_age = 5, max_age = 100)}
}
