% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anthro_age_distribution.R
\name{plot_anthro_age_distribution}
\alias{plot_anthro_age_distribution}
\title{Plot Age Distribution of Nutritional Status of Children}
\usage{
plot_anthro_age_distribution(
  df,
  index,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe with anthropometric data that has already been processed
and standardized by healthyr::format_nut_health_results function.}

\item{index}{Inputs a character value specifying which anthropometric index or result to plot.
Options include wfhz, hfaz, wfaz, mfaz, muac, or cgam.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot object.
}
\description{
Plot Age Distribution of Nutritional Status of Children
}
\examples{
\dontrun{plot_anthro_age_distribution(df, index = "wfhz",
file_path = "myplots/wfhz_age_distribution.jpg")}
}
