% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlogram.R
\name{plot_correlogram}
\alias{plot_correlogram}
\title{Plot Correlogram}
\usage{
plot_correlogram(
  df,
  numeric_cols,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe.}

\item{numeric_cols}{Inputs a character vector of specifying the column names
of the numeric variables to use.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function wraps the GGally package to facilitate making correlation plots
to visually assess the relationships between several numeric variables.
}
\examples{
\dontrun{plot_correlogram(df, numeric_cols = c("fcs_score", "rcsi_score", "hhs_score"))}
}
