% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_distribution.R
\name{plot_cumulative_distribution}
\alias{plot_cumulative_distribution}
\title{Plot Cumulative Distribution}
\usage{
plot_cumulative_distribution(
  df,
  index,
  flags,
  grouping = NULL,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{A data frame with MUAC data, each row is 1 child.}

\item{index}{Inputs a character value for which anthropometric index to plot. Use 'wfhz' for weight-for-height,
'hfaz' for height-for-age, 'wfaz' for weight-for-age, 'mfaz' for MUAC-for-age, or 'muac' for MUAC.}

\item{flags}{Inputs a character value of 'yes' or 'no'. Yes specifies to include flagged values in the plot. No specifies
to exclude flagged values in the plot.}

\item{grouping}{Inputs an optional character value to facet the plot by.#' @return Returns a ggplot cumulative distribution plot of MUAC measurements.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\description{
Plots a cumulatative distribution of MUAC or an anthropometric index.
}
\examples{
\dontrun{plot_cumulative_distribution(df, index = "muac", flags = "yes", grouping = "enum")}
}
