% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_quality.R
\name{plot_cumulative_quality}
\alias{plot_cumulative_quality}
\title{Plot Cumulative Quality}
\usage{
plot_cumulative_quality(df, quality_indicator)
}
\arguments{
\item{df}{Inputs a dataframe which has an been processed and standandardized by
healthyr::format_nut_health_indicators function.}

\item{quality_indicator}{Inputs a character value specifying the quality indicator to plot.
Options include sd and dps.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function creates a plot of a numeric quality indicator over time so an analyst can
observe if an enumerator's work is improving over the course of data collection, primarily
with anthropometric data.
}
\examples{
\dontrun{plot_cumulative_quality(df, quality_indicator = sd)}
}
