% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_domain_distribution.R
\name{plot_domain_distribution}
\alias{plot_domain_distribution}
\title{Plot Select Multiple Across Domains}
\usage{
plot_domain_distribution(
  df,
  domain_vars,
  response_vars,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe.}

\item{domain_vars}{Inputs a character vector of column names for domain categories. Each domain
should have its own column coded as 1/0.}

\item{response_vars}{Inputs a character vector of column names for response categories for
the select multiple question coded as 1/0.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This functions plots a count distribution of responses for a select multiple question
and shows their responses distributed against some domain categorization of those
response. E.g A health barriers question with several response options, however those
response options can be grouped by barriers related to availability, accessibility, or
quality of services. Each response option should only belong to one domain category.
}
\examples{
\dontrun{plot_domain_distribution(df, domain_vars = c("availability",
"accessibility", "quality"), response_vars = c("barrier.distance",
"barrier.nostaff", "barrier.costtransport", "barrier.rudestaff"))}
}
