% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_bargraph.R
\name{plot_results_bargraph}
\alias{plot_results_bargraph}
\title{Plot Bar Graph Single Variable}
\usage{
plot_results_bargraph(
  results_df,
  col_results,
  var_to_plot,
  var_to_group,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{results_df}{Inputs a dataframe of results produced from the
healthyr::analyse_survey_results function.}

\item{col_results}{Inputs a character vector of column names that are formatted as
results with CIs in square brackets.}

\item{var_to_plot}{Inputs a character value specifying the variable to plot.}

\item{var_to_group}{Inputs a character value specifying the variable to group by, for
grouped bar charts.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function takes a single variable result from a dataframe of results produced by
healthyr::analyse_survey_results and creates a bar chart with confidence intervals.
If specified can also produce grouped bar charts. Intended for use on already
aggregated results, uses stat = "identity".
}
\examples{
\dontrun{plot_results_bargraph(results_df, col_results = c(names(df[,-1]),
 var_to_plot = "gam_wfhz_noflag", var_to_group = "admin2") ))}
}
