% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridge_distribution.R
\name{plot_ridge_distribution}
\alias{plot_ridge_distribution}
\title{Plot Ridge Distributions}
\usage{
plot_ridge_distribution(
  df,
  numeric_cols,
  name_groups = NULL,
  name_units = NULL,
  grouping = NULL,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe.}

\item{numeric_cols}{Inputs a character vector specifying column names of the numeric
variables to plot.}

\item{name_groups}{Inputs an optional character value specifying the names of the groups.}

\item{name_units}{Inputs an optional character value specifying the names of the units.}

\item{grouping}{Inputs an optional character value specifying the name of a column by which
to facet the plot.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
Creates a distribution plot of several numeric variables side by side for comparison.
}
\examples{
\dontrun{plot_ridge_distribution(df, numeric_cols = c("fcs_cereal", "fcs_dairy", ...),
name_groups = "Food Groups", name_units = "Days", grouping = "enum")}
}
