% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_select_multiple.R
\name{plot_select_multiple}
\alias{plot_select_multiple}
\title{Plot Select Multiple Results}
\usage{
plot_select_multiple(
  results_df,
  select_multiple_cols,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{results_df}{Inputs a dataframe of results from the healthyr::analyse_survey_results
function.}

\item{select_multiple_cols}{Inputs a character vector of column names for the select
multiple question to plot.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function creates a plot results for a select multiple indicator that has been
aggregated already through the healthyr::analyse_survey_results function.
}
\examples{
\dontrun{plot_select_multiple(results_df = myresults,
select_multiple_cols = c("barrier.1", "barrier.2", "barrier.3"))}
}
