% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vars_by_group.R
\name{plot_vars_by_group}
\alias{plot_vars_by_group}
\title{Plot Variables by Group}
\usage{
plot_vars_by_group(
  df,
  vars,
  grouping,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{df}{Inputs a dataframe.}

\item{vars}{Inputs a character vector of column names to plot by group.}

\item{grouping}{Inputs a character value specifying the column by which to
facet the plot.}

\item{file_path}{Inputs an optional character value specifying the file path location to save
an image of the plot. Recommended to specify as a .jpg or .png file.}

\item{wdth}{Inputs an optional numeric value for the width of the image.}

\item{hght}{Inputs an optional numeric value fo the height of the image.}

\item{title_name}{Inputs an optional character value for the title of the plot.}
}
\value{
Returns a ggplot2 object.
}
\description{
This function creates bar charts showing counts of responses by group.
}
\examples{
\dontrun{plot_vars_by_group(df, vars = c("availability", "accessibility", "quality"),
grouping = "unmet_need")}
}
