% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gartner_magic_chart.R
\name{plt_gartner_magic_chart}
\alias{plt_gartner_magic_chart}
\title{Gartner Magic Chart Plotting of two continuous variables}
\usage{
plt_gartner_magic_chart(.data, .x_col, .y_col, .y_lab, .x_lab, .plt_title)
}
\arguments{
\item{.data}{The data set you want to plot}

\item{.x_col}{The x-axis for the plot}

\item{.y_col}{The y-axis for the plot}

\item{.y_lab}{The y-axis label}

\item{.x_lab}{The x-axis label}

\item{.plt_title}{The title of the plot}
}
\value{
A ggplot plot
}
\description{
Plot a Gartner Magic Chart of two continuous variabls
}
\details{
\itemize{
\item Supply a data frame with at least two continuous variables to plot against
each other
}
}
\examples{
library(dplyr)
library(ggplot2)
library(purrr)
library(tibble)
plt_gartner_magic_chart(
  .data = tibble(x = rnorm(100, 0, 1), y = rnorm(100,0,1))
  , .x_col = x
  , .y_col = y
  , .x_lab = "los"
  , .y_lab = "ra"
  , .plt_title = "tst"
)

}
