% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_anthro1.R
\docType{data}
\name{proc_anthro1}
\alias{proc_anthro1}
\title{SMART Survey Anthropometric Data - Processed}
\format{
A data frame with 651 rows and 99 variables:
\describe{
\item{hh_id}{Unique household survey identifier.}
\item{sex}{Sex of the child}
\item{child_present}{Yes (1)/No (0) if the child is present at the time of the household interview.}
\item{dob_known}{Yes (1)/No (0) if the date of birth of the child is known.}
\item{dob}{Date of birth of the child, if known}
\item{age_months}{Age in months of the child, from date of birth or estimate from a local events calendar.}
\item{weight}{Weight in kilogram of the child}
\item{measure}{If the child was measured by Length (1) or by Height (2)}
\item{height}{The length or height of the child in cm.}
\item{muac}{The mid-upper arm circumference of the child in mm.}
\item{edema}{Yes (1)/No (0) if the child had bilateral pitting oedema.}
\item{oedema}{Yes (1)/No (0) if the team supervisor confirmed the bilateral putting oedema.
Standardized variable name. This variable should be used for oedema in analysis.}
\item{county}{The administrative name of the county where the survey took place.}
\item{date_dc}{The date of data collection}
\item{enum}{The survey team or enumerator ID}
\item{cluster}{The cluster id}
\item{date_dc_date}{Date of data collection, as a date type variable.}
\item{date_dc_month}{Month of data collection.}
\item{date_dc_day}{Day of the month of data collection.}
\item{date_dc_year}{Year of data collection.}
\item{date_dc_char}{Date of data collection, as a character type variable.}
\item{dob_date}{Date of birth, as a date type variable.}
\item{dob_month}{Month of date of birth.}
\item{dob_day}{Day of date of birth.}
\item{dob_year}{Year of date of birth.}
\item{dob_char}{Date of birth, as a character type variable.}
\item{age_days}{Age in days based on date of birth or estimated age in months.}
\item{age_years}{Age in completed years.}
\item{age_group}{Age group in groups of 5 years.}
\item{age_proxy}{Age group if the child is <2 years or >2 years of age.}
\item{measured}{If the child was measured by Length (1) or by Height (2)}
\item{wfhz}{Weight-for-height z-score}
\item{sam_wfhz}{Yes (1)/No (0) if child has severe acute malnutrition by weight for height z-score.}
\item{mam_wfhz}{Yes (1)/No (0) if child has moderate acute malnutrition by weight for height z-score.}
\item{gam_wfhz}{Yes (1)/No (0) if child has global acute malnutrition by weight for height z-score.}
\item{hfaz}{Height-for-age z-score}
\item{severe_stunting}{Yes (1)/No (0) if child has severe stunting by height-for-age z-score of < -3 and/or nutritional oedema.}
\item{moderate_stunting}{Yes (1)/No (0) if child has moderate stunting by height-for-age z-score >= -3 and < -2.}
\item{global_stunting}{Yes (1)/No (0) if child has global stunting by height-for-age z-score of < -2 and/or nutritional oedema.}
\item{wfaz}{Weight-for-age z-score}
\item{severe_underweight}{Yes (1)/No (0) if child has severe underweight by weight-for-age z-score of < -3.}
\item{moderate_underweight}{Yes (1)/No (0) if child has moderate underweight by weight-for-age z-score >= -3 and < -2.}
\item{global_underweight}{Yes (1)/No (0) if child has global underweight by weight-for-age z-score of < -2.}
\item{mfaz}{MUAC-for-age z-score}
\item{severe_mfaz}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC-for-age z-score of < -3.}
\item{moderate_mfaz}{Yes (1)/No (0) if child has moderate acute malnutrition by MUAC-for-age z-score >= -3 and < -2.}
\item{global_mfaz}{Yes (1)/No (0) if child has global acute malnutrition by MUAC-for-age z-score of < -2.}
\item{sam_muac}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC < 115mm and/or nutritional oedema}
\item{mam_muac}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC >=115 and < 125mm}
\item{gam_muac}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC < 125mm and/or nutritional oedema}
\item{flag_missing_sex}{Yes (1)/No (0) Data quality flag if sex is missing/NA.}
\item{flag_missing_age_years}{Yes (1)/No (0)  Data quality flag if age_years is missing/NA.}
\item{mean_wfhz}{Mean weight-for-height z-score for the specified grouping or strata.}
\item{sd_wfhz}{Standard deviation for weight-for-height z-score for the specified grouping or strata.}
\item{wfhz_smart_flag}{Yes (1)/No (0) Data quality flag if wfh z-score exceeds +/- 3 from the mean wfh z-score.}
\item{wfhz_who_flag}{Yes (1)/No (0) Data quality flag if wfh z-score exceeds +/- 5. }
\item{wfhz_noflag}{Weight-for-height z-score with flagged values set to missing.}
\item{mean_wfhz_noflag}{Mean weight-for-height z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{sd_wfhz_noflag}{Standard deviation for weight-for-height z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{gam_wfhz_noflag}{Yes (1)/No (0) if child has global acute malnutrition by weight for height z-score, with SMART flagged values excluded.}
\item{mam_wfhz_noflag}{Yes (1)/No (0) if child has moderate acute malnutrition by weight for height z-score, with SMART flagged values excluded.}
\item{sam_wfhz_noflag}{Yes (1)/No (0) if child has severe acute malnutrition by weight for height z-score, with SMART flagged values excluded.}
\item{mean_hfaz}{Mean height-for-age z-score for the specified grouping or strata.}
\item{sd_hfaz}{Standard deviation for height-for-age z-score for the specified grouping or strata.}
\item{hfaz_smart_flag}{Yes (1)/No (0) Data quality flag if hfa z-score exceeds +/- 3 from the mean hfa z-score.}
\item{hfaz_who_flag}{Yes (1)/No (0) Data quality flag if hfa z-score exceeds +/- 5. }
\item{hfaz_noflag}{Height-for-age z-score with flagged values set to missing.}
\item{mean_hfaz_noflag}{Mean height-for-age z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{sd_hfaz_noflag}{Standard deviation for height-for-age z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{global_stunting_noflag}{Yes (1)/No (0) if child has global stunting by height for age z-score, with SMART flagged values excluded.}
\item{moderate_stunting_noflag}{Yes (1)/No (0) if child has moderate stunting by height for age z-score, with SMART flagged values excluded.}
\item{severe_stunting_noflag}{Yes (1)/No (0) if child has severe stunting by height for age z-score, with SMART flagged values excluded.}
\item{mean_wfaz}{Mean weight-for-age z-score for the specified grouping or strata.}
\item{sd_wfaz}{Standard deviation for weight-for-age z-score for the specified grouping or strata.}
\item{wfaz_smart_flag}{Yes (1)/No (0) Data quality flag if wfa z-score exceeds +/- 3 from the mean wfa z-score.}
\item{wfaz_who_flag}{Yes (1)/No (0) Data quality flag if wfa z-score exceeds +/- 5. }
\item{wfaz_noflag}{Weight-for-age z-score with flagged values set to missing.}
\item{mean_wfaz_noflag}{Mean weight-for-age z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{sd_wfaz_noflag}{Standard deviation for weight-for-age z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{global_underweight_noflag}{Yes (1)/No (0) if child has global stunting by weight for age z-score, with SMART flagged values excluded.}
\item{moderate_underweight_noflag}{Yes (1)/No (0) if child has moderate stunting by weight for age z-score, with SMART flagged values excluded.}
\item{severe_underweight_noflag}{Yes (1)/No (0) if child has severe stunting by weight for age z-score, with SMART flagged values excluded.}
\item{mean_mfaz}{Mean MUAC-for-age z-score for the specified grouping or strata.}
\item{sd_mfaz}{Standard deviation for MUAC-for-age z-score for the specified grouping or strata.}
\item{mfaz_smart_flag}{Yes (1)/No (0) Data quality flag if mfa z-score exceeds +/- 3 from the mean wfa z-score.}
\item{mfaz_who_flag}{Yes (1)/No (0) Data quality flag if mfa z-score exceeds +/- 5. }
\item{mfaz_noflag}{MUAC-for-age z-score with flagged values set to missing.}
\item{mean_mfaz_noflag}{Mean MUAC-for-age z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{sd_mfaz_noflag}{Standard deviation for MUAC-for-age z-score for the specified grouping or strata, calculated without SMART flagged values.}
\item{global_mfaz_noflag}{Yes (1)/No (0) if child has global acute malnutrition by MUAC for age z-score, with SMART flagged values excluded.}
\item{moderate_mfaz_noflag}{Yes (1)/No (0) if child has moderate acute malnutrition by MUAC for age z-score, with SMART flagged values excluded.}
\item{severe_mfaz_noflag}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC for age z-score, with SMART flagged values excluded.}
\item{muac_flag}{Yes (1)/No (0) Data quality flag if MUAC is an extreme value <8cm or >22cm for a child under-5 years of age.}
\item{muac_noflag}{MUAC measurement with extreme values excluded/set to missing.}
\item{gam_muac_noflag}{Yes (1)/No (0) if child has global acute malnutrition by MUAC, with extreme values excluded.}
\item{mam_muac_noflag}{Yes (1)/No (0) if child has moderate acute malnutrition by MUAC, with extreme values excluded.}
\item{sam_muac_noflag}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC, with extreme values excluded.}
\item{c_gam}{Yes (1)/No (0) if child has global acute malnutrition by MUAC, wfh z-score < -2, or presence of nutritional oedema,
with SMART flagged and extreme values excluded.}
\item{c_sam}{Yes (1)/No (0) if child has severe acute malnutrition by MUAC, wfh z-score < -2, or presence of nutritional oedema,
with SMART flagged and extreme values excluded.}

}
}
\source{
SMART Survey August 2019
}
\usage{
proc_anthro1
}
\description{
A dataset containing household anthropometric data for children under-5 years of age.
This dataset has been processed by the healthyr functions, including flag_anthro_issues.
}
\examples{
data(proc_anthro1)
}
\keyword{datasets}
