% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_healthbarriers_fsl3.R
\docType{data}
\name{proc_healthbarriers_fsl3}
\alias{proc_healthbarriers_fsl3}
\title{Multi-Sectoral Needs Assessment FSL and Health Barriers - Processed}
\format{
A data frame with 950 rows and 69 variables:
\describe{
\item{hh_id}{Unique household survey identifier.}
\item{strata}{Survey strata.}
\item{svy_weights}{Household level survey weights.}
\item{population_group}{Population group of the household.}
\item{region}{Region admin level 1}
\item{governorate}{Governate admin level 2}
\item{district}{District admin level 3}
\item{fcs_cereal}{Number of days consumed cereals or tubers in the past 7 days.}
\item{fcs_legumes}{Number of days consumed nuts, seeds, legumes in the past 7 days.}
\item{fcs_dairy}{Number of days consumed milk or dairy products in the past 7 days.}
\item{fcs_meat}{Number of days consumed meats, fish, eggs in the past 7 days.}
\item{fcs_veg}{Number of days consumed vegetables in the past 7 days.}
\item{fcs_fruit}{Number of days consumed fruits in the past 7 days.}
\item{fcs_oil}{Number of days consumed oils or fats in the past 7 days.}
\item{fcs_sugar}{Number of days consumed sugar or sweets in the past 7 days.}
\item{spices_condiments}{Number of days consumed any condiments like salt, tea, coffee in the past 7 days.}
\item{barrier.none}{Household self-reported health barriers - None (1)/(0)}
\item{barrier.no_facility_nearby}{Household self-reported health barriers - No facility nearby (1)/(0)}
\item{barrier.hours_inconvenient}{Household self-reported health barriers - Service hours inconvenient (1)/(0)}
\item{barrier.service_unavailable}{Household self-reported health barriers - Service unavailable (1)/(0)}
\item{barrier.long_wait}{Household self-reported health barriers - Long waiting time for service (1)/(0)}
\item{barrier.cost_consultation}{Household self-reported health barriers - Cannot afford cost of consultation (1)/(0)}
\item{barrier.cost_treatment}{Household self-reported health barriers - Cannot afford cost of treatment (1)/(0)}
\item{barrier.cost_transportation}{Household self-reported health barriers - Cannot afford cost of transportation (1)/(0)}
\item{barrier.medication_issues}{Household self-reported health barriers - Inappropriate medication (1)/(0)}
\item{barrier.distance}{Household self-reported health barriers - Far distance to health facility (1)/(0)}
\item{barrier.disability}{Household self-reported health barriers - Limited or no access for disabled at facility (1)/(0)}
\item{barrier.no_transportation}{Household self-reported health barriers - No transport to facility (1)/(0)}
\item{barrier.insecurity_at_facility}{Household self-reported health barriers - Insecurity at facility (1)/(0)}
\item{barrier.insecurity_traveling}{Household self-reported health barriers - Insecurity en route to facility. (1)/(0)}
\item{barrier.fear_covid}{Household self-reported health barriers - Afraid of getting COVID-19 (1)/(0)}
\item{barrier.no_trained_staff}{Household self-reported health barriers - No trained staff at facility (1)/(0)}
\item{barrier.insufficient_staff}{Household self-reported health barriers - Not enough staff at facility (1)/(0)}
\item{barrier.no_female_staff}{Household self-reported health barriers - No female staff at facility (1)/(0)}
\item{barrier.distrust}{Household self-reported health barriers - Lack of trust with facility staff (1)/(0)}
\item{barrier.no_free_time}{Household self-reported health barriers - No free time to go to facility (1)/(0)}
\item{barrier.language_communication}{Household self-reported health barriers - Facility staff don't speak my language (1)/(0)}
\item{barrier.lack_docs}{Household self-reported health barriers - Lack legal documentation (1)/(0)}
\item{barrier.employer_prevented}{Household self-reported health barriers - Employer prevented going to facility (1)/(0)}
\item{barrier.other}{Household self-reported health barriers - Other barrier (1)/(0)}
\item{barrier.dk}{Household self-reported health barriers - Don't know (1)/(0)}
\item{barrier.decline}{Household self-reported health barriers - Declined/Refused to answer (1)/(0)}
\item{health_barrier.none}{Health Barrier Domain - No Barriers (1)/0}
\item{health_barrier.phys_access}{Health Barrier Domain - Physical Accessibility (1)/0}
\item{health_barrier.insecurity}{Health Barrier Domain - Insecurity (1)/0}
\item{health_barrier.fin_access}{Health Barrier Domain - Financial Accessibility (1)/0}
\item{health_barrier.legal_access}{Health Barrier Domain - Legal Accessibility (1)/0}
\item{health_barrier.availability}{Health Barrier Domain - Availability (1)/0}
\item{health_barrier.quality}{Health Barrier Domain - Quality (1)/0}
\item{health_barrier.cultural}{Health Barrier Domain - Cultural (1)/0}
\item{health_barrier.other}{Health Barrier Domain - Other (1)/0}
\item{health_barrier.dknr}{Health Barrier Domain - Don't know/No response (1)/0}
\item{fcs_weight_cereal1}{Weighted value for Food Consumption Score. fcs_cereals x 2}
\item{fcs_weight_legume2}{Weighted value for Food Consumption Score. fcs_legumes x 3}
\item{fcs_weight_dairy3}{Weighted value for Food Consumption Score. fcs_dairy x 4}
\item{fcs_weight_meat4}{Weighted value for Food Consumption Score. fcs_meat x 4}
\item{fcs_weight_veg5}{Weighted value for Food Consumption Score. fcs_veg x 1}
\item{fcs_weight_fruit6}{Weighted value for Food Consumption Score. fcs_fruit x 1}
\item{fcs_weight_oil7}{Weighted value for Food Consumption Score. fcs_oil x 0.5}
\item{fcs_weight_sugar8}{Weighted value for Food Consumption Score. fcs_sugar x 0.5}
\item{fcs_score}{Food Consumption Score value. Sum of fcs_weight variables.}
\item{fcs_cat}{Food Consumption Score categorization. May be reported with normal or alternative thresholds.
Normal thresholds are Poor (<21), Borderline (>=21 & <35), and Acceptable (>=35). Alternative thresholds may
be used if high oil and sugar consumption is observed, and are Poor (<28), Borderline (>=28 & <42), and
Acceptable (>=42).}
\item{flag_above7_fcs}{Data quality flag if any FCS questions that report greater than 7 days food consumption in the past 7 days.}
\item{flag_meat_cereal_ratio}{Data quality flag if more days of meat consumption are reported than days of cereal consumption in the past 7 days.}
\item{flag_zero_fcs}{Data quality flag if FCS score of 0.}
\item{flag_all_fcs}{Data quality flag if all FCS values are 0 or 7.}
\item{flag_low_fcs}{Data quality flag if FCS score < 5 is reported.}
\item{flag_high_fcs}{Data quality flag if FCS score > 60 is reported.}
\item{flag_fcs_extreme}{Data quality flag if an extreme FCS score of <3 or >60 is reported.}

}
}
\source{
Multi-Sectoral Needs Assessment October 2021
}
\usage{
proc_healthbarriers_fsl3
}
\description{
A dataset containing household surveys from an MSNA. Has Food Consumption Scores,
and Health Barriers indicators. This data has been processed by healthyr functions.
}
\examples{
data(proc_healthbarriers_fsl3)
}
\keyword{datasets}
