% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-files.R
\docType{data}
\name{px_cc_mapping}
\alias{px_cc_mapping}
\title{Procedure to Condition Code Mapping file}
\format{
A data frame with 79721 rows and 5 variables
}
\usage{
data(px_cc_mapping)
}
\description{
Procedure to Condition Code Mapping file
}
\seealso{
Other AHRQ: 
\code{\link{dx_cc_mapping}}
}
\concept{AHRQ}
\keyword{datasets}
