% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_helper_data.R
\name{recode_helper_data}
\alias{recode_helper_data}
\title{Recode Helper Data}
\usage{
recode_helper_data(df, column, old_val, new_val)
}
\arguments{
\item{df}{The dataframe}

\item{column}{The column name to change}

\item{old_val}{The old value}

\item{new_val}{The new value}
}
\value{
A recoded dataframe.
}
\description{
Recode Helper Data
}
