% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_helper_map.R
\name{recode_helper_map}
\alias{recode_helper_map}
\title{Recode Helper Map}
\usage{
recode_helper_map(df, column, old_val, new_val)
}
\arguments{
\item{df}{The dataframe}

\item{column}{A character value of the column name to recode.}

\item{old_val}{The old value to recode.}

\item{new_val}{The new value to recode.}
}
\value{
Returns a re-coded dataframe for this single variable.
}
\description{
Helper function to update the map dataframe based on user console inputs
for the recoding steps
}
