% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_mortality_current_census.R
\name{reformat_mortality_current_census}
\alias{reformat_mortality_current_census}
\title{Reformatting Mortality Data for Current Census Survey Method}
\usage{
reformat_mortality_current_census(df)
}
\arguments{
\item{df}{Inputs a dataframe where the column names have been standardized already from the format_mortality_current_census function.}
}
\value{
Returns a reformatted standardized long mortality dataset.
}
\description{
This function is intended as a helper function and to be called from within the format_mortality_current_census function.
It will check the formatting of each of the variables, and if needed prompt the user for recoding certain variables.
}
\examples{
\dontrun{reformat_mortality_current_census(df)}

}
