% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_nut_health_indicators.R
\name{reformat_nut_health_indicators}
\alias{reformat_nut_health_indicators}
\title{Reformat Nutrition Health Indicators}
\usage{
reformat_nut_health_indicators(
  df,
  health_barriers = NULL,
  lcs_variables = NULL,
  livelihood_variables = NULL,
  value_map
)
}
\arguments{
\item{df}{Inputs a data frame directly from the format_nut_health_indicators function.}

\item{health_barriers}{Inputs a character vector of column names for self-reported health barriers.}

\item{lcs_variables}{Inputs a character vector of column names for livelihood coping strategies.}

\item{livelihood_variables}{Inputs a character vector of column names specifying all the livelihood activities
for which income is received. These should be numerical columns with estimated income per livelihood activity.
As per MSNA core indicator bank 2023.}

\item{value_map}{Inputs a value map which is generated from other heatlhyr functions to facilitate recoding.}
}
\value{
Returns a dataframe with standardized values for specific columns, that can then be used to calculate composite or other indicators.
}
\description{
#' Function to reformat and standardize the values of columns for nutrition and health analysis. This function is intended to be a helper function
and to be called from within the format_nut_health_indicators function after column names and preliminary checks have been made. This
function will check the values for each relevant column and in some cases prompt the user for recoding inputs. This function covers anthropometric columns,
Washington Group Short Set indicators, infant and young child feeding indicators, food security outcome indicators, and a few other MSNA relevant health indicators.
}
\examples{
\dontrun{reformat_nut_health_indicators(df = myhealthdata, health_barriers =
c("healthbarrier_colname_1", "healthbarrier_colname_2", ...), lcs_variables =
c("lcs_colname1", "lcs_colname2", ...))}

}
