% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_iycf_monitoring_dashboard.R
\name{run_iycf_monitoring_dashboard}
\alias{run_iycf_monitoring_dashboard}
\title{Run IYCF Monitoring Dashboard}
\usage{
run_iycf_monitoring_dashboard(
  df,
  grouping_var = NULL,
  filter_var1 = NULL,
  filter_var2 = NULL
)
}
\arguments{
\item{df}{A dataframe with IYCF data that has been formatted by healthry::format_nut_health_indicators function.}

\item{grouping_var}{A character value naming the column to group the results by.}

\item{filter_var1}{A character value naming the first column to filter the data by.}

\item{filter_var2}{A character value naming an additional column to filter the data by.}
}
\value{
Opens a shiny dashboard.
}
\description{
Runs a local shiny dashboard to explore data quality for IYCF outcome indicators
}
