% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_to_excel}
\alias{save_to_excel}
\title{Save a file to Excel}
\usage{
save_to_excel(.data, .file_name)
}
\arguments{
\item{.data}{The tibble/data.frame that you want to save as an \code{.xlsx} file.}

\item{.file_name}{the name you want to give to the file.}
}
\value{
A saved excel file
}
\description{
Save a tibble/data.frame to an excel \code{.xlsx} file. The file will automatically
with a save_dtime in the format of 20201109_132416 for November 11th, 2020
at 1:24:16PM.
}
\details{
\itemize{
\item Requires a tibble/data.frame to be passed to it.
}
}
\examples{
\donttest{
library(tibble)
tibble::tibble("x" = runif(10, 0, 1), "y" = runif(10, 0, 1)) \%>\%
  save_to_excel(.file_name = "test_file")
}

}
\author{
Steven P. Sanderson II, MPH
}
