% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pallette-color-blind.R
\name{scale_color_colorblind}
\alias{scale_color_colorblind}
\title{Provide Colorblind Compliant Colors}
\usage{
scale_color_colorblind(..., theme = "td")
}
\description{
Provide Colorblind Compliant Colors
}
