% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_ae_admission}
\alias{scotland_ae_admission}
\title{Scotland A&E Admission Outcomes}
\format{
A data frame with 86,678 rows and 6 variables:
\describe{
\item{hospital_code}{Hospital code}
\item{hospital_name}{Hospital name}
\item{date}{Date}
\item{admission}{Admission outcome}
\item{age}{Age group}
\item{attendances}{Total number of attendances falling into each category}
...
}
}
\source{
\url{https://publichealthscotland.scot/healthcare-system/urgent-and-unscheduled-care/accident-and-emergency/downloads-and-open-data/our-downloads/}
}
\usage{
scotland_ae_admission
}
\description{
A dataset containing Scotland hospital-level admission outcomes stratified by age group (from 2018-01 to 2025-02)
}
\keyword{datasets}
