% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_ae_attendance}
\alias{scotland_ae_attendance}
\title{Scotland A&E attendances}
\format{
A data frame with 17,844 rows and 6 variables:
\describe{
\item{hospital_code}{Hospital code}
\item{hospital_name}{Hospital name}
\item{date}{Date}
\item{attendances}{Number of total attendances}
\item{within_4_hours}{Number of attendances lasting less than 4 hours}
\item{over_4_hours}{Number of attendances lasting more than 4 hours}
...
}
}
\source{
\url{https://publichealthscotland.scot/healthcare-system/urgent-and-unscheduled-care/accident-and-emergency/downloads-and-open-data/our-downloads/}
}
\usage{
scotland_ae_attendance
}
\description{
A dataset of hospital-level A&E attendances and waiting times for Scotland (from 2007-07 to 2025-02).
}
\keyword{datasets}
