% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_beds}
\alias{scotland_beds}
\title{Bed occupancy/availability in Scotland}
\format{
A data frame with 11,029 rows and 6 variables:
\describe{
\item{hb19_code}{Scottish Health Board code}
\item{date}{Date, representing the start of the quarter}
\item{specialty}{Name of the specialty}
\item{average_number_available_staffed_beds}{Daily average number of available staffed beds}
\item{average_number_occupied_beds}{Daily average number of occupied beds}
\item{percent_occupied_beds}{Percentage (\%) of daily occupancy of beds}
...
}
}
\source{
\url{https://www.opendata.nhs.scot}
}
\usage{
scotland_beds
}
\description{
Beds by Board of Treatment and Specialty, 2016-2024.
}
\keyword{datasets}
