% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_care_at_home}
\alias{scotland_care_at_home}
\title{Care at Home in Scotland}
\format{
A data frame containing 800 rows and 4 columns.
\describe{
\item{ltla24_code}{2024 Lower tier local authority code}
\item{ltla24_name}{2024 Lower tier local authority name}
\item{date}{Date}
\item{care_at_home_per_1000}{Number of people (per 1,000 population) receiving care at home}
}
}
\source{
\url{https://publichealthscotland.scot/publications/care-at-home-statistics-for-scotland/care-at-home-statistics-for-scotland-support-and-services-funded-by-health-and-social-care-partnerships-in-scotland-20232024/dashboard/}
}
\usage{
scotland_care_at_home
}
\description{
Scotland Health-Board level data for the number of people (per 1,000 population) supported by care at home (from 2018-Q1 to 2024-Q1).
}
\keyword{datasets}
