% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_delayed_discharge_hb}
\alias{scotland_delayed_discharge_hb}
\title{Delayed Discharge Bed Days by Health Board in Scotland}
\format{
A data frame with 24,750 rows and 6 variables:
\describe{
\item{hb_code}{Scottish Health Board code}
\item{date}{Date}
\item{age_group}{Age grouping is calculated as at the person's ready for discharge date}
\item{delay_reason}{Reason for delay indicates the principal reason grouping for a person's delay at the end of the reporting month}
\item{num_delayed_bed_days}{The total number of delayed bed days}
\item{average_daily_delayed_beds}{The average daily number of delayed beds is calculated by dividing the total number of delayed discharge bed days in the month by the number of days in the calendar month}
...
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
scotland_delayed_discharge_hb
}
\description{
A dataset containing Delayed Discharge Bed Days by Scottish Health Board.
}
\keyword{datasets}
