% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_rtt}
\alias{scotland_rtt}
\title{Consultant-led Referral to Treatment (RTT) in Scotland}
\format{
A data frame with 141 rows and 4 variables:
\describe{
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
scotland_rtt
}
\description{
A dataset containing RTT data for Scotland.
}
\keyword{datasets}
