% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_underdoctored_areas}
\alias{scotland_underdoctored_areas}
\title{Underdoctored areas in Scotland}
\format{
A data frame with 5 variables:
\describe{
\item{ltla21_code}{Local Authority code (2021)}
\item{total_gp}{Number of full-time equivalent GPs}
\item{total_patients}{Number of registered patients}
\item{patients_per_gp}{Number of patients per GP}
\item{underdoctored_decile}{Decile of number of patients per GP (10 = highest number of patients)}
}
}
\usage{
scotland_underdoctored_areas
}
\description{
A dataset containing the ratio of patients to GP workforce, by Local Authority
}
\keyword{datasets}
