% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_non_date_cols.R
\name{select_non_date_cols}
\alias{select_non_date_cols}
\title{Select Non-Date Columns}
\usage{
select_non_date_cols(df)
}
\arguments{
\item{df}{A dataframe object.}
}
\value{
Returns a dataframe object without any date columns.
}
\description{
Helper function to select only non-Date columns. will test each column for
Date properties or formatting, and if present will remove that column.
}
