% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_non_numeric_cols.R
\name{select_non_numeric_cols}
\alias{select_non_numeric_cols}
\title{Select Non-Numeric Columns
Helper function to select only non-numeric columns in a dataset.
It will also exclude character columns that have only numeric values.#'}
\usage{
select_non_numeric_cols(df)
}
\arguments{
\item{df}{A dataframe object.}
}
\value{
Returns a dataframe object with only non-numeric columns selected.
}
\description{
Select Non-Numeric Columns
Helper function to select only non-numeric columns in a dataset.
It will also exclude character columns that have only numeric values.#'
}
