% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_non_other_columns.R
\name{select_non_other_columns}
\alias{select_non_other_columns}
\title{Select Non Other Columns}
\usage{
select_non_other_columns(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A dataframe object removing any columns with the suffixes "_other" or "_autre".
}
\description{
Select all columns that do not end in "_other" or "_autre"
}
