% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_string_split.R
\name{sql_left}
\alias{sql_left}
\title{Use SQL LEFT type function}
\usage{
sql_left(text, num_char)
}
\arguments{
\item{text}{A piece of text/string to be manipulated}

\item{num_char}{How many characters do you want to grab}
}
\description{
Perform a sql LEFT() type function on a piece of text
}
\details{
\itemize{
\item You must supply data that you want to manipulate.
}
}
\examples{

sql_left("text", 3)

}
