% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_mortality_period.R
\name{subset_mortality_period}
\alias{subset_mortality_period}
\title{Subset Mortality Period}
\usage{
subset_mortality_period(df, start_date, end_date)
}
\arguments{
\item{df}{Inputs a dataframe that has already been standardized by the format_current_census_mortality function.}

\item{start_date}{Inputs a character value of a date of the start of the recall period, in a format like YYYY-MM-DD.}

\item{end_date}{Inputs a character value of a date of the end of the recall period, in a format like YYYY-MM-DD.}
}
\value{
Returns a subsetted mortality dataset.
}
\description{
Intended to be used on a standardized mortality dataset which additionally captured dates of
different demographic events such as joins, lefts, births and deaths. It will dynamically subset the data
between a specified start and end of the recall period, and either include or exclude people, and calculate
their estimated person time observed.
}
\examples{
\dontrun{subset_mortality_period(df = mymortalitydata, start_date = "2020-04-24", end_date = "2021-05-15")}
}
