% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_values_by_group.R
\name{summarize_value_by_group}
\alias{summarize_value_by_group}
\title{Summarize Value by Group}
\usage{
summarize_value_by_group(df, grouping, val)
}
\arguments{
\item{df}{Inputs a dataframe}

\item{grouping}{Inputs a character value of the column name by which to group the results.}

\item{val}{Inputs a character value of the value that you want to summarize across groups and columns.}
}
\value{
Returns a dataframe of a summary table counting the number of occurences of the value in each column, by group.
}
\description{
Creates a summary table of the number of times a specified value appears in each dataframe column, by group.
For summarizing missing values, use instead healthyr::summarize_NA_by_group
}
\examples{
\dontrun{summarize_value_by_group(df = mydata, grouping = "enum", val = "dontknow")}
}
