% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_median_excess_plt.R
\name{ts_median_excess_plt}
\alias{ts_median_excess_plt}
\title{Create a plot showing the excess +/- of the median value}
\usage{
ts_median_excess_plt(
  .data,
  .date_col,
  .value_col,
  .x_axis,
  .ggplt_group_var,
  .secondary_grp_var,
  ...
)
}
\arguments{
\item{.data}{The data that is being analyzed, data must be a tibble/data.frame}

\item{.date_col}{The column of the tibble that holds the date}

\item{.value_col}{The column that holds the value of interest}

\item{.x_axis}{What is the be the x-axis, day, week, etc.}

\item{.ggplt_group_var}{What is the resulting ggplot grouped on, e.g. Year}

\item{.secondary_grp_var}{Typically the same as the x-axis, may not work as expected otherwise}

\item{...}{Same as the .ggplot_group_var and .secondary_grp_var}
}
\description{
Plot out the excess +/- of the median value grouped by certain time parameters
}
\details{
\itemize{
\item Supply data that you want to view and you will see the excess +/- of the median values
over a specified time series tibble
}
}
\examples{
library(timetk)
library(dplyr)

ts_median_excess_plt(
  .data = m4_daily
  , .date_col = date
  , .value_col = value
  , .x_axis = wk
  , .ggplt_group_var = yr
  , .secondary_group_var = wk
  , yr
  , wk
)


}
