% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_plots.R
\name{ts_plot_readmit_rate}
\alias{ts_plot_readmit_rate}
\title{Plot Readmit Rate}
\usage{
ts_plot_readmit_rate(.data, .date_col, .value_col, .by_grouping)
}
\arguments{
\item{.data}{The data you need to pass}

\item{.date_col}{The date column}

\item{.value_col}{The value column}

\item{.by_grouping}{How you want the data summarized - "sec", "min", "hour", "day", "week", "month", "quarter" or "year"}
}
\value{
A timetk time series plot that is interactive
}
\description{
Plot ALOS - average length of stay
}
\details{
\itemize{
\item Expects a tibble with a date time column and a value column
\item Uses timetk for underlying sumarization and plot
\item If .by_grouping is missing it will default to "day"
}
}
\author{
Steven P. Sanderson II, MPH
}
