% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans-funcs.R
\name{user_item_tbl}
\alias{user_item_tbl}
\title{OPPE CPOE K-Means}
\usage{
user_item_tbl(.data, .user_input, .item_input)
}
\arguments{
\item{.data}{The data that you want to transform}

\item{.user_input}{The column that is going to be the row (user)}

\item{.item_input}{The column that is going to be the column (item)}
}
\value{
A aggregated/normalized user item tibble
}
\description{
Takes in a data.frame/tibble and transforms it into an aggregated/normalized
user-item tibble of proportions. The user will need to input the parameters
for the rows/user and the columns/items.
}
\details{
This function should be used before using a k-mean model. This is
commonly referred to as a user item matrix because "users" tend to be on the
rows and "items" (e.g. orders) on the columns.
}
\examples{
user_item_tbl()

}
\author{
Steven P. Sanderson II, MPH
}
