% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_ae_performance}
\alias{wales_ae_performance}
\title{Emergency department performance statistics}
\format{
A data frame with 73,863 rows and 5 variables:
\describe{
\item{hospital_code}{Hospital code}
\item{hospital_department}{Hospital or department name}
\item{date}{Date}
\item{performance_measure}{Performance measure}
\item{performance_target}{Performance target}
\item{data}{Data}
}
}
\source{
\url{https://statswales.gov.wales/}
}
\usage{
wales_ae_performance
}
\description{
A dataset containing performance against waiting times targets by hospital.
}
\keyword{datasets}
