% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_ambulance_waiting_times}
\alias{wales_ambulance_waiting_times}
\title{Wales ambulance waiting times - (2022)}
\format{
A data frame with 288 rows and 9 variables:
\describe{
\item{date}{Date}
\item{hb_code}{Local health board code}
\item{hb}{Name of local health board}
\item{red_calls}{Number of red calls}
\item{red_calls_resulting_in_an_emergency_response_at_the_scene}{Number of red call resulting in an emergency response at the scene}
\item{amber_calls}{Number of amber calls}
\item{green_calls}{Number of green calls}
\item{red_calls_resulting_in_an_emergency_response_at_the_scene_within_8_minutes}{Number of red call resulting in an emergency response at the scene within 8 minutes}
\item{red_calls_percent_of_emergency_responses_arriving_at_the_scene_within_8_minutes}{Percentage of emergency responses to red calls arriving at the scene within 8 minutes}

...
}
}
\source{
\url{https://statswales.gov.wales/}
}
\usage{
wales_ambulance_waiting_times
}
\description{
A dataset containing Welsh emergency ambulance calls and responses to red
calls, by local health boards (HB) and month
}
\keyword{datasets}
