% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_crime_severity}
\alias{wales_crime_severity}
\title{Crime severity score in Wales}
\format{
A data frame containing 2 columns.
\describe{
\item{ltla23_code}{Local Authority code}
\item{crime_severity_score}{Crime severity score (higher = more severe)}
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/crimeseverityscoreexperimentalstatistics}
}
\usage{
wales_crime_severity
}
\description{
Crime severity score is based on sentencing data, community order and fines.
Higher scores mean greater severity within an area.
}
\keyword{datasets}
