% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_delayed_discharge_reasons}
\alias{wales_delayed_discharge_reasons}
\title{Delayed discharge reasons for Wales, 2023 April - 2025 March}
\format{
A data frame with 5,220 rows and 5 variables:
\describe{
\item{lhb22_code}{Health Board code}
\item{lhb22_name}{Health Board name}
\item{date}{Date in the format year-month-day}
\item{delay_reason}{The reason for the delayed discharge}
\item{delay_count}{The number of delays}
...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Performance/pathway-of-care-delays/pathwayofcaredelays-by-localhealthboardprovider-date}
}
\usage{
wales_delayed_discharge_reasons
}
\description{
A dataset containing delayed discharge data stratified by local health board and the reason for delay.
}
\keyword{datasets}
