% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_delayed_discharges}
\alias{wales_delayed_discharges}
\title{Welsh Delayed Discharges}
\format{
A data frame with 50,400 rows and 5 variables:
\describe{
\item{lhb22_name}{Local Health Board name}
\item{lhb22_code}{Local Health Board code}
\item{date}{Date}
\item{delay_reason}{Reason for the delayed discharge}
\item{delay_count}{Count of the delayed discharge}
...
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
wales_delayed_discharges
}
\description{
A dataset containing the count and reason of delayed discharges in Welsh
Health Boards from April 2023 to March 2025.
}
\keyword{datasets}
