% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_disability_21}
\alias{wales_disability_21}
\title{Disability by LTLA for Wales, 2021}
\format{
A data frame:
\describe{
\item{ltla21_code}{LTLA code}
\item{total_residents}{Total residents in the local authority}
\item{disability_level}{Disabled or Not disabled under the Equality Act}
\item{n}{number of people}
\item{prop}{proportion of people}
...
}
}
\source{
\url{https://www.nomisweb.co.uk/sources/census_2021_bulk}
}
\usage{
wales_disability_21
}
\description{
A dataset containing LTLA data on numbers disabled under the Equality Act.
}
\keyword{datasets}
