% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_health_board_critical_general_acute_beds}
\alias{wales_health_board_critical_general_acute_beds}
\title{Wales health board bed availability - (2022 - )}
\format{
A data frame with 1,595 rows and 7 variables:
\describe{
\item{date}{Date}
\item{health_board_code}{Local health board code}
\item{health_board_name}{Name of local health board}
\item{specialism}{Name of specialism}
\item{daily_beds_available}{Average number of daily beds available}
\item{daily_beds_occupied}{Average number of daily beds occucpied}
\item{beds_occupancy_rate}{Bed occupancy rate}

...
}
}
\source{
\url{https://statswales.gov.wales/}
}
\usage{
wales_health_board_critical_general_acute_beds
}
\description{
A dataset containing Welsh monthly NHS beds data by health board, measure,
and specialty
}
\keyword{datasets}
