% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_homelessness}
\alias{wales_homelessness}
\title{Homelessness in Wales}
\format{
A data frame with 3 variables:
\describe{
\item{ltla21_code}{Local Authority code}
\item{Homeless or threatened with homelessness per 1,000}{Households assessed as homeless per 1,000 households in the Local Authority}
\item{Households in temporary accommodation per 1,000}{Households in temporary accommodation per 1,000 households in the Local Authority}
}
}
\source{
\url{https://statswales.gov.wales/}
}
\usage{
wales_homelessness
}
\description{
A dataset containing statistics on households assessed as homeless and
households in temporary accommodation, by Local Authority.
}
\keyword{datasets}
