% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_underdoctored_areas}
\alias{wales_underdoctored_areas}
\title{Underdoctored areas in Wales}
\format{
A data frame with 4 variables:
\describe{
\item{ltla24_code}{Local Authority code (2024)}
\item{total_patients}{Number of registered patients}
\item{total_gp}{Number of GPs}
\item{patients_per_gp}{Number of patients per GP}
}
}
\usage{
wales_underdoctored_areas
}
\description{
A dataset containing the ratio of patients to GP workforce, by Local Authority
}
\keyword{datasets}
