% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2_funs.R
\name{xml2_import_health_data}
\alias{xml2_import_health_data}
\title{read health data}
\usage{
xml2_import_health_data(path, type = "xml")
}
\arguments{
\item{path}{chr path to xml file}

\item{type}{chr file type; default 'xml' and only supports xml for now}
}
\value{
xml document; see \code{xml2::read_xml()}
}
\description{
read health data
}
