% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{dx_cc_mapping}
\alias{dx_cc_mapping}
\title{Diagnosis to Condition Code Mapping file}
\format{
A data frame with 86852 rows and 5 variables
}
\usage{
data(dx_cc_mapping)
}
\description{
A dataset containing the Diagnosis Code to AHRQ Condition Code Mapping
that is used in helping to define service lines for inpatient discharges.
}
\details{
\itemize{
\item CC_Code. DX_1, DX_2, ..., DX_n
\item CC_Desc. DX_1 = Conduction disorders, DX_n = description
\item ICD_Ver_Flag. ICD Version 10 or 9
\item ICDCode. ICD-9 ro ICD-10 Code
\item Diagnosis. Long QT Syndrome
}
}
\keyword{datasets}
