% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_critical_general_acute_beds}
\alias{england_critical_general_acute_beds}
\title{Critical care and General & Acute Beds}
\format{
A data frame with 959 rows and 20 variables:
\describe{
\item{nhs_trust22_code}{NHS Trust (organisational) code}
\item{date}{Date}
\item{general_acute_beds_available}{"G&A beds available"}
\item{general_acute_beds_occupied}{"G&A beds occupied"}
\item{general_acute_beds_occupancy_rate}{"G&A occupancy rate"}
\item{adult_general_acute_beds_available}{"Adult G&A beds available"}
\item{adult_general_acute_beds_occupied}{"Adult G&A beds occupied"}
\item{adult_general_acute_beds_occupancy_rate}{"Adult G&A occupancy rate"}
\item{paediatric_general_acute_beds_available}{"Paediatric G&A beds available"}
\item{paediatric_general_acute_beds_occupied}{"Paediatric G&A beds occupied"}
\item{paediatric_general_acute_beds_occupancy_rate}{"Paediatric G&A occupancy rate"}
\item{adult_critical_care_beds_available}{"Adult critical care beds available"}
\item{adult_critical_care_beds_occupied}{"Adult critical care beds occupied"}
\item{adult_critical_care_occupancy_rate}{"Adult critical care occupancy rate"}
\item{paediatric_intensive_cared_beds_available}{"Paediatric intensive care beds available"}
\item{paediatric_intensive_cared_beds_occupied}{"Paediatric intensive care beds occupied"}
\item{paediatric_intensive_cared_occupancy_rate}{"Paediatric intensive care occupancy rate"}
\item{neonatal_intensive_care_bed_avaialble}{"Neonatal intensive care beds available"}
\item{neonatal_intensive_care_bed_occupied}{"Neonatal intensive care beds occupied"}
\item{neonatal_intensive_care_occupancy_rate}{"Neonatal intensive care occupancy rate"}
...
}
}
\source{
\url{https://www.england.nhs.uk/}
}
\usage{
england_critical_general_acute_beds
}
\description{
A dataset containing NHS Trust level critical care and general and acute bed
availability information. This is taken from the urgent and emergency care
daily situation reports. Bed numbers are daily figures averaged over the
month.
}
\keyword{datasets}
