% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_health_index}
\alias{england_health_index}
\title{England Health Index (ONS)}
\format{
A data frame with 2,149 rows and 3 variables:
\describe{
\item{ltla21_code}{2021 Lower tier local authority code}
\item{year}{Year}
\item{overall_score}{The overall health index score}
...
}
}
\source{
\url{https://www.ons.gov.uk/}
}
\usage{
england_health_index
}
\description{
A dataset containing overall health index scores for the English lower tier
local authorities from 2015-2021.
}
\keyword{datasets}
