% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_health_index_subdomains}
\alias{england_health_index_subdomains}
\title{England Health Index (ONS) - subdomains}
\format{
A data frame with 2,149 rows and 16 variables:
\describe{
\item{ltla21_code}{2021 Lower tier local authority code}
\item{year}{Year}
\item{Difficulties in daily life \link{Pe}}{Difficulties in daily life \link{Pe}}
\item{Mental health \link{Pe}}{Mental health \link{Pe}}
\item{Mortality \link{Pe}}{Mortality \link{Pe}}
\item{Personal well-being \link{Pe}}{Personal well-being \link{Pe}}
\item{Physical health conditions \link{Pe}}{Physical health conditions \link{Pe}}
\item{Behavioural risk factors \link{L}}{Behavioural risk factors \link{L}}
\item{Children and young people \link{L}}{Children and young people \link{L}}
\item{Physiological risk factors \link{L}}{Physiological risk factors \link{L}}
\item{Protective measures \link{L}}{Protective measures \link{L}}
\item{Access to green space \link{Pl}}{Access to green space \link{Pl}}
\item{Access to services \link{Pl}}{Access to services \link{Pl}}
\item{Crime \link{Pl}}{Crime \link{Pl}}
\item{Economic and working conditions \link{Pl}}{Economic and working conditions \link{Pl}}
\item{Living conditions \link{Pl}}{Living conditions \link{Pl}}
...
}
}
\source{
\url{https://www.ons.gov.uk/}
}
\usage{
england_health_index_subdomains
}
\description{
A dataset containing overall health index scores for the English lower tier
local authorities from 2015-2021.
}
\keyword{datasets}
