% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_homelessness}
\alias{england_homelessness}
\title{Statutory homelessness in England}
\format{
A data frame with 3 variables:
\describe{
\item{ltla23_code}{Local Authority code}
\item{\verb{Households assessed as homeless per (000s)}}{Households assessed as homeless per 1,000 households in the Local Authority}
\item{\verb{Households in temporary accommodation per 1,000}}{Households in temporary accommodation per 1,000 households in the Local Authority}
}
}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/live-tables-on-homelessness}
}
\usage{
england_homelessness
}
\description{
A dataset containing statistics on households assessed as homeless and
households in temporary accommodation, by Local Authority.
}
\keyword{datasets}
