% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ni_unpaid_care_21}
\alias{ni_unpaid_care_21}
\title{Provision of unpaid care (delayed discharge) in Northern Ireland}
\format{
A data frame with 11 rows and 8 variables:
\describe{
\item{ltla21_name}{Local Government Districts name}
\item{ltla21_code}{Local Government Districts code}
\item{total_population}{Population aged over 5 years}
\item{total_unpaid_carers}{Total number of unpaid carers}
\item{age_5_to_14_years}{Unpaid carers aged between 5 and 14}
\item{age_15_to_39_years}{Unpaid carers aged between 15 and 39}
\item{age_40_to_64_years}{Unpaid carers aged between 40 and 64}
\item{age_65_over_years}{Unpaid carers aged over 65}
...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
ni_unpaid_care_21
}
\description{
A dataset containing unpaid care data in Local Government Districts in NI
from 2021 census.
}
\keyword{datasets}
