% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gartner_magic_chart.R
\name{plt_gartner_magic_chart}
\alias{plt_gartner_magic_chart}
\title{Gartner Magic Chart Plotting of two continuous variables}
\usage{
plt_gartner_magic_chart(.data, .x_col, .y_col)
}
\arguments{
\item{.data}{The data set you want to plot}

\item{.x_col}{The x-axis for the plot}

\item{.y_col}{The y-axis for the plot}
}
\value{
A ggplot plot
}
\description{
Plot a Gartner Magic Chart of two continuous variabls
}
\details{
\itemize{
\item Supply a data frame with at least two continuous variables to plot against
each other
}
}
\examples{
library(tidyverse)
library(purrr)
x <- rnorm(180, mean = 0, sd = 1)
y <- rnorm(180, mean = 0, sd = 1)
df <- data.frame(x,y)
df \%>\%
  plt_gartner_magic_chart(
    .x_axis = x
    , .y_axis = y
  )

}
