% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{px_cc_mapping}
\alias{px_cc_mapping}
\title{Procedure to Condition Code Mapping file}
\format{
A data frame with 79721 rows and 5 variables
}
\usage{
data(px_cc_mapping)
}
\description{
A dataset containing the Procedure Code to AHRQ Condition Code Mapping
that is used in helping to define servce lines for inpatient discharges.
}
\details{
\itemize{
\item CC_Code. PX_1, PX_2, ..., PX_n
\item CC_Desc. PX_1 = Genitourinary incontinence procedures
\item ICD_Ver_Flag. 10 or 9
\item ICDCode. ICD-9 or ICD-10 Code
\item Procedure. Inject Implant Urethra
}
}
\keyword{datasets}
