% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_delayed_discharge_ltla}
\alias{scotland_delayed_discharge_ltla}
\title{Delayed Discharge Bed Days by Council Area in Scotland}
\format{
A data frame with 41,310 rows and 6 variables:
\describe{
\item{ltla_code}{Scottish Council Area code}
\item{date}{Date}
\item{age_group}{Age grouping is calculated as at the person's ready for discharge date}
\item{delay_reason}{Reason for delay indicates the principal reason grouping for a person's delay at the end of the reporting month}
\item{num_delayed_bed_days}{The total number of delayed bed days}
\item{average_daily_delayed_beds}{The average daily number of delayed beds is calculated by dividing the total number of delayed discharge bed days in the month by the number of days in the calendar month}
...
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
scotland_delayed_discharge_ltla
}
\description{
A dataset containing Delayed Discharge Bed Days by Scottish Council Area
}
\keyword{datasets}
