% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotland_rtt_hb}
\alias{scotland_rtt_hb}
\title{Consultant-led Referral to Treatment (RTT) by Scottish Health Board}
\format{
A data frame with 2,091 rows and 6 variables:
\describe{
\item{hb19_code}{Scottish Health Board code}
\item{date}{Date}
\item{year}{Year}
\item{month}{Month}
\item{specialty}{Specialty}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
}
}
\source{
\url{https://www.opendata.nhs.scot/}
}
\usage{
scotland_rtt_hb
}
\description{
A dataset containing RTT data for Scottish Health Boards.
}
\keyword{datasets}
