% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_string_split.R
\name{sql_mid}
\alias{sql_mid}
\title{Use SQL MID type function}
\usage{
sql_mid(text, start_num, num_char)
}
\arguments{
\item{text}{A piece of text/string to be manipulated}

\item{start_num}{What place to start at}

\item{num_char}{How many characters do you want to grab}
}
\description{
Perform a SQL SUBSTRING type function
}
\details{
\itemize{
\item You must supply data that you want to manipulate.
}
}
\examples{

sql_mid("this is some text", 6, 2)

}
