% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_plots.R
\name{ts_plot_alos}
\alias{ts_plot_alos}
\title{Plot ALOS - Average length of stay}
\usage{
ts_plot_alos(.data, .date_col, .value_col, .by_grouping)
}
\arguments{
\item{.data}{The data you need to pass}

\item{.date_col}{The date column}

\item{.value_col}{The value column}

\item{.by_grouping}{How you want the data summarized - "sec", "min", "hour", "day", "week", "month", "quarter" or "year"}
}
\value{
A timetk time series plot that is interactive
}
\description{
Plot ALOS - average length of stay
}
\details{
\itemize{
\item Expects a tibble with a date time column and a value column
\item Uses timetk for underlying sumarization and plot
\item If .by_grouping is missing it will default to "day"
}
}
\examples{
library(dplyr)
library(purrr)
library(timetk)

ts_tbl <- tk_make_timeseries(start = "2019-01-01", by = "day", length_out = "1 year 6 months")
values <- arima.sim(model= list(order = c(0, 1, 0)), n=547, mean=1,sd=5)
df_tbl <- tibble(x = ts_tbl, y = values) \%>\% set_names("Date","Values")

ts_plot_alos(.data = df_tbl, .date_col = Date, .value_col = Values, .by = "month")

}
\author{
Steven P. Sanderson II, MPH
}
