% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_hospitals_critical_general_acute_beds}
\alias{wales_hospitals_critical_general_acute_beds}
\title{Wales hospital bed availability - (2022)}
\format{
A data frame with 1610 rows and 21 variables:
\describe{
\item{date}{Date}
\item{hospital_code}{Local health board code}
\item{hospital_name}{Name of local health board}
\item{all_specialties_average_daily_beds_available}{All specialties beds available}
\item{all_specialties_average_daily_beds_occupied}{All specialties beds occupied}
\item{all_specialties_beds_occupancy_rate}{All specialties occupancy rate}
\item{surgical_acute_average_daily_beds_available}{Surgical acute beds available}
\item{surgical_acute_average_daily_beds_occupied}{Surgical acute beds occupied}
\item{surgical_acute_beds_occupancy_rate}{Surgical acute occupancy rate}
\item{medical_acute_average_daily_beds_available}{Medical acute beds available}
\item{medical_acute_average_daily_beds_occupied}{Medical acute beds occupied}
\item{medical_acute_beds_occupancy_rate}{Medical acute occupancy rate}
\item{mental_illness_average_daily_beds_available}{Mental illness beds available}
\item{mental_illness_average_daily_beds_occupied}{Mental illness beds occupied}
\item{mental_illness_beds_occupancy_rate}{Mental illness occupancy rate}
\item{geriatric_medicine_average_daily_beds_occupied}{Geriatric beds available}
\item{geriatric_medicine_average_daily_beds_available}{Geriatric beds occupied}
\item{geriatric_medicine_beds_occupancy_rate}{Geriatric occupancy rate}
\item{maternity_average_daily_beds_available}{Maternity beds available}
\item{maternity_average_daily_beds_occupied}{Maternity beds occupied}
\item{maternity_beds_occupancy_rate}{Maternity occupancy rate}

...
}
}
\source{
\url{https://statswales.gov.wales/}
}
\usage{
wales_hospitals_critical_general_acute_beds
}
\description{
A dataset containing Welsh monthly NHS beds data by hospital site, measure,
and specialty
}
\keyword{datasets}
