% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wales_rtt_lhb}
\alias{wales_rtt_lhb}
\title{Consultant-led Referral to Treatment (RTT) by Local Health Board}
\format{
A data frame with 1,260 rows and 6 variables:
\describe{
\item{lhb22_code}{Local Health Board code}
\item{lhb22_name}{Local Health Board name}
\item{date}{Date}
\item{pathway_stage}{Stage of the treatment pathway}
\item{waits_over_18_weeks}{Total waiting > 18 weeks}
\item{waits_over_53_weeks}{Total waiting > 53 weeks}
}
}
\source{
\url{https://statswales.gov.wales/}
}
\usage{
wales_rtt_lhb
}
\description{
A dataset containing RTT data for Wales.
}
\keyword{datasets}
